Gibbon.Compiler

Description

The compiler pipeline, assembled from several passes.

Compiler entrypoints

compile :: Config -> FilePath -> IO ()

Compiler entrypoint, given a full configuration and a list of files to process, do the thing.

compileCmd :: [String] -> IO ()

Command line version of the compiler entrypoint. Parses command line arguments given as string inputs. This also allows us to run conveniently from within GHCI. For example:

 compileCmd $ words $ " -r -p -v5 examples/test11c_funrec.gib "

Configuration options and parsing

data Config

Overall configuration of the compiler, as determined by command line arguments and possible environment variables.

Constructors

Config

Fields

input :: Input

mode :: Mode

How to run, which backend.

benchInput :: Maybe FilePath

What packed, binary .gpkd file to use as input.

arrayInput :: Maybe FilePath

What array file to use as input.

verbosity :: Int

Debugging output, equivalent to DEBUG env var.

cc :: String

C compiler to use

optc :: String

Options to the C compiler

cfile :: Maybe FilePath

Optional override to destination .c file.

exefile :: Maybe FilePath

Optional override to destination binary file.

backend :: Backend

Compilation backend used

dynflags :: DynFlags

srcFile :: Maybe FilePath

The file being compiled by Gibbon.

data Mode

How far to run the compiler/interpreter.

Constructors

ToParse
Parse and then stop

ToC
Compile to C

ToExe
Compile to C then build a binary.

RunExe
Compile to executable then run.

Interp2
Interp late in the compiler pipeline.

Interp1
Interp early.

ToMPL
Compile to SML (mlton dialect)

ToMPLExe
Compile to SML & compile with MPL

RunMPL
Compile to SML & compile with MPL & run

Bench Var
Benchmark a particular function applied to the packed data within an input file.

BenchInput FilePath
Hardcode the input file to the benchmark in the C code.

data Input

What input format to expect on disk.

Constructors

Haskell

SExpr

Unspecified

configParser :: Parser Config

configWithArgs :: Parser (Config, [FilePath])

Parse configuration as well as file arguments.

defaultConfig :: Config

Some other helper fns

compileAndRunExe :: Config -> FilePath -> IO String

Compile and run the generated code if appropriate