Gibbon.DynFlags

Description

Flags à la GHC

Documentation

data DynFlags

Constructors

DynFlags

Fields

generalFlags :: Set GeneralFlag

debugFlags :: Set DebugFlag

data GeneralFlag

Constructors

Opt_Gibbon1
Set Opt_No_RemoveCopies & Opt_BigInfiniteRegions

Opt_Gibbon2
Set Opt_RemoveCopies & Opt_InfiniteRegions

Opt_RemoveCopies
Calls to copy functions are converted to indirections

Opt_No_RemoveCopies
Unset Opt_RemoveCopies

Opt_InfiniteRegions
Use infinite regions

Opt_BigInfiniteRegions
Use big infinite regions

Opt_BenchPrint
Should the benchamrked function have its output printed?

Opt_Packed
Use packed representation

Opt_Pointer
Use pointer representation

Opt_BumpAlloc
Use bump-pointer allocation if using the non-packed backend

Opt_Warnc
Show warnings from the C compiler

Opt_DisableGC
Don't run the the garbage collector (used by Codegen).

Opt_No_PureAnnot
Don't use pure annotations (a GCC optimization)

Opt_Fusion
Enable fusion.

Opt_Parallel
Fork/join parallelism.

Opt_RegionOnSpawn
Allocate into fresh regions for every spawn, not steal.

Opt_GhcTc
Typecheck with GHC before compiling with Gibbon.

Opt_RelativeOffsets
Enable relative offsets.

Opt_CountParRegions
Count and print the number of regions allocated for parallelism.

Opt_CountAllRegions
Count and print the number of all the regions allocated.

Opt_RtsDebug
Compile the RTS in debugging mode.

Opt_PrintGcStats
Record and print GC statistics.

Opt_GenGc
Use the new generational GC.

Opt_NoEagerPromote
Disable eager promotion.

Opt_SimpleWriteBarrier
Disables eliminate-indirection-chains optimization.

data DebugFlag

Exactly like GHC's ddump flags.

Constructors

Opt_D_Dump_Repair

Opt_D_Dump_ParAlloc

Opt_D_DumpToFile

Opt_D_Dump_Hs

defaultDynFlags :: DynFlags

dynflagsParser :: Parser DynFlags

gopt :: GeneralFlag -> DynFlags -> Bool

Test whether a GeneralFlag is set

gopt_set :: GeneralFlag -> DynFlags -> DynFlags

dopt :: DebugFlag -> DynFlags -> Bool

dopt_set :: DebugFlag -> DynFlags -> DynFlags