Gibbon.Interp

Description

Documentation

class Interp a where

Pure Gibbon programs, at any stage of compilation, should always be evaluatable to a unique value. The only side effects are timing.

Minimal complete definition

interpProg

Methods

interpProg :: RunConfig -> a -> IO (Value, ByteString)

interpNoLogs :: RunConfig -> a -> String

Interpret while ignoring timing constructs, and dropping the corresponding output to stdout.

interpWithStdout :: RunConfig -> a -> IO (String, [String])

Interpret and produce a "log" of output lines, as well as a final, printed result. The output lines include timing information.

data Store

A store is an address space full of buffers.

Constructors

Store (Map Var Buffer)

insertIntoStore :: MonadState Store m => Var -> Buffer -> m ()

lookupInStore :: MonadState Store m => Var -> m (Maybe Buffer)

data Buffer

Constructors

Buffer (Seq SerializedVal)

emptyBuffer :: Buffer

insertAtBuffer :: Int -> SerializedVal -> Buffer -> Buffer

Insert a value at a particular index in the buffer

data SerializedVal

Constructors

SerTag Word8 DataCon

SerInt Int

SerBool Int

data Value

It's a first order language with simple values.

Constructors

VInt Int

VBool Bool

VDict (Map Value Value)

VProd [Value]

VPacked DataCon [Value]

VLoc

Fields

bufID :: Var

offset :: Int

VCursor
Cursor are a pointer into the Store plus an offset into the Buffer.

Fields

bufID :: Var

offset :: Int

type Log = Builder

type ValEnv = Map Var Value

execAndPrint :: Interp (Prog ex) => RunConfig -> Prog ex -> IO ()

deserialize :: Out ty => DDefs ty -> Seq SerializedVal -> Value

Code to read a final answer back out.

strToInt :: String -> Int

lookup3 :: (Eq k, Show k, Show a, Show b) => k -> [(k, a, b)] -> (k, a, b)

Orphan instances

Out Word8

Instance details

Methods

docPrec :: Int -> Word8 -> Doc

doc :: Word8 -> Doc

docList :: [Word8] -> Doc

Out a => Out (Seq a)

Instance details

Methods

docPrec :: Int -> Seq a -> Doc

doc :: Seq a -> Doc

docList :: [Seq a] -> Doc