Gibbon.L0.Specialize2

Documentation

l0ToL1 :: Prog0 -> PassM Prog1

toL1 :: Prog0 -> Prog1

type MonoM a = StateT MonoState PassM a

data MonoState

Constructors

MonoState

Fields

mono_funs_todo :: Map (Var, [Ty0]) Var

mono_funs_done :: Map (Var, [Ty0]) Var

mono_lams :: Map (Var, [Ty0]) Var

mono_dcons :: Map (TyCon, [Ty0]) Var

emptyMonoState :: MonoState

extendFuns :: (Var, [Ty0]) -> Var -> MonoState -> MonoState

extendLambdas :: (Var, [Ty0]) -> Var -> MonoState -> MonoState

extendDatacons :: (TyCon, [Ty0]) -> Var -> MonoState -> MonoState

getLambdaObls :: Var -> MonoState -> Map Var [Ty0]

monomorphize :: Prog0 -> PassM Prog0

assertLambdasMonomorphized :: Monad m => MonoState -> m ()

assertSameLength :: (Out a, Out b, Monad m) => String -> [a] -> [b] -> m ()

collectMonoObls :: DDefs0 -> Env2 Ty0 -> Set Var -> L Exp0 -> MonoM (L Exp0)

Collect monomorphization obligations.

monoLambdas :: L Exp0 -> MonoM (L Exp0)

Create monomorphic versions of lambdas bound in this expression. This does not float out the lambda definitions.

purgePolyFuns :: Prog0 -> Prog0

Remove all polymorphic functions and datatypes from a program. monoLambdas already gets rid of polymorphic mono_lams.

isMonoFun :: FunDef0 -> Bool

purgePolyDDefs :: Prog0 -> Prog0

updateTyCons :: MonoState -> Prog0 -> Prog0

updateTyConsExp :: DDefs0 -> MonoState -> L Exp0 -> L Exp0

updateTyConsTy :: DDefs0 -> MonoState -> Ty0 -> Ty0

Update TyCons if an appropriate monomorphization obligation exists.

type SpecM a = StateT SpecState PassM a

type FunRef = Var

data SpecState

Constructors

SpecState

Fields

sp_funs_todo :: Map (Var, [FunRef]) Var

sp_fundefs :: FunDefs0

spec :: Prog0 -> PassM Prog0

specFun :: DDefs0 -> Env2 Ty0 -> Var -> [FunRef] -> FunDef0 -> SpecM ()

specExp :: DDefs0 -> Env2 Ty0 -> L Exp0 -> SpecM (L Exp0)