Gibbon.L0.Syntax

Description

The higher-ordered surface language.

extended expressions

type Exp0 = E0 () Ty0

Extended expressions, L0.

data E0Ext loc dec

The extension that turns L1 into L0.

Constructors

LambdaE (Var, dec) (L (E0 loc dec))
Variable tagged with type Can't attach haddocks to data constructor arguments with < GHC 8.4.2

PolyAppE (L (E0 loc dec)) (L (E0 loc dec))

data PProg

Monomorphic program

Constructors

PProg

Fields

pddefs :: PDDefs Ty0

pfundefs :: PFDefs Ty0 (L Exp0)

pvardefs :: PVDefs Ty0 (L Exp0)

pmainExp :: Maybe (L Exp0)

data MProg

Constructors

MProg

Fields

ddefs :: DDefs Ty0

fundefs :: FunDefs0

vardefs :: VarDefs Ty0 (L Exp0)

mainExp :: Maybe (L Exp0)

curried functions vs regular functions

type CurFun = VarDef Ty0 Exp

we now have curried functions and curried calls curried functions are these variable defns but curried calls vs function calls are PolyAppE vs AppE

type L0Fun = FunDef0

Monomorphized functions

type CCall = Exp

type FCall = Exp

types and operation on types

data Ty0

Constructors

IntTy

BoolTy

TyVar Var

ProdTy [Ty0]

SymDictTy Ty0

ArrowTy Ty0 Ty0

PackedTy TyCon [TyVar]

ListTy Ty0

data Scheme a

Constructors

ForAll [TyVar] a

initFunEnv :: PFDefs Ty0 Exp -> FunEnv Ty0

initVarEnv :: PVDefs Ty0 Exp -> Map Var Ty0

arrIn :: Ty0 -> Ty0

arrOut :: Ty0 -> Ty0

typeFromScheme :: Scheme a -> a

TEMP : as it is already defined in Common but need to update branch

variable definitions

data VarDef a ex

Variable definitions

Monomorphic version

Constructors

VarDef

Fields

varName :: Var

varTy :: a

varBody :: ex

type VarDefs a ex = Map Var (VarDef a ex)

polymorphic versions of definitions (functions,vars,ddefs)

data PVDef a ex

Constructors

PVDef

Fields

vName :: Var

vTy :: Scheme a

vBody :: ex

type PVDefs a ex = Map Var (PVDef a ex)

data PFDef a ex

Polymorphic program

Constructors

PFDef

Fields

fName :: Var

fArg :: Var

fTy :: Scheme a

the type will be a ForAll

fBody :: ex

type PFDefs a ex = Map Var (PFDef a ex)

for now, using a specialized FunDef for L0 theoretically these should disappear after monomorphization this enables the FunDefs to have type schemes

data PDDef a

Constructors

PDDef
Polymorphic data constructors

Fields

dName :: Var

dCons :: [(DataCon, [(IsBoxed, Scheme a)])]

type PDDefs a = Map Var (PDDef a)

for now, using a specialized DDef for L0 this enables the DDefs to have type variables

old function definitions

type FunDefs0 = Map Var FunDef0

type FunDef0 = FunDef (L Exp0)