Gibbon.L1.Examples

Documentation

treeTy :: Ty1

treeDD :: DDefs (UrTy ())

mkAdd1Prog :: Exp1 -> Maybe (Exp1, Ty1) -> Prog1

mkAdd1Fun :: Exp1 -> FunDef1

add1Prog :: Prog1

The basic form of the add1 program where recursions happen immediately as arguments to the data-constructor.

exadd1Bod :: Exp1

exadd1BodLetLeft :: Exp1

exadd1BodLetRight :: Exp1

A more challenging case where recursions are performed right-to-left

add1ProgLetLeft :: Prog1

Add1 program with let bindings, recurring in left-to-right order.

add1ProgLetRight :: Prog1

Add1 program with let bindings, recurring in right-to-left order.

add1ProgChallenge :: Prog1

An even more challenging case where there is an (apparent) data dependency where x2 depends on y2.

add1ProgSharing :: Prog1

This program is a challenge because a packed value flows to two destinations.