Gibbon.L1.Interp

Description

Interpreter for L1

Documentation

interpProg :: Interp () e => RunConfig -> Prog e -> IO ((), Value e, ByteString)

Interpret a program, including printing timings to the screen. The returned bytestring contains that printed timing info.

interp :: forall e l d s. (Show l, Ord l, NFData l, Out l, Show d, NFData d, Out d, Ord d, Ord (e l d), NFData (e l d), InterpExt s (PreExp e l d) (e l d)) => RunConfig -> ValEnv (PreExp e l d) -> DDefs (TyOf (PreExp e l d)) -> FunDefs (PreExp e l d) -> PreExp e l d -> InterpM s (PreExp e l d) (Value (PreExp e l d))

applyPrim :: (Show ty, Ord l, Out l, Show l, Show d, Out d, Ord d, Ord (e l d), Out (e l d), Show (e l d)) => RunConfig -> Prim ty -> [Value (PreExp e l d)] -> InterpM s (PreExp e l d) (Value (PreExp e l d))

Orphan instances

Interp () Exp1

Instance details

Methods

gInterpExp :: RunConfig -> ValEnv Exp1 -> DDefs (TyOf Exp1) -> FunDefs Exp1 -> Exp1 -> InterpM () Exp1 (Value Exp1)

InterpProg () Exp1

Instance details

Methods

gInterpProg :: () -> RunConfig -> Prog Exp1 -> IO ((), Value Exp1, ByteString)

gInterpNoLogs :: () -> RunConfig -> Prog Exp1 -> String

gInterpWithStdout :: () -> RunConfig -> Prog Exp1 -> IO (String, [String])

InterpExt () Exp1 (E1Ext () Ty1)

Instance details

Methods

gInterpExt :: RunConfig -> ValEnv Exp1 -> DDefs (TyOf Exp1) -> FunDefs Exp1 -> E1Ext () Ty1 -> InterpM () Exp1 (Value Exp1)