Gibbon.L1.Syntax

Description

The source language for recursive tree traversals. This is a first-order language for the "closed world" scenario: not integrating with a functional host language, but rather genarating C code like a DSL.

Core types specific to L1

type Prog1 = Prog (L Exp1)

An L1 program.

type FunDef1 = FunDef (L Exp1)

Function definition used in L1 programs.

type FunDefs1 = FunDefs (L Exp1)

type DDef1 = DDef Ty1

type DDefs1 = DDefs Ty1

Datatypes

type Exp1 = PreExp NoExt () Ty1

A convenient, default instantiation of the L1 expression type.

type Ty1 = UrTy ()

The type rperesentation used in L1.

data NoExt l d

An uninhabidited type indicating that the base grammar is not extended with any additional constructs.

module Gibbon.Language

Orphan instances

FunctionTy Ty1

Instance details

Associated Types

type ArrowTy Ty1 :: *

Methods

inTys :: ArrowTy Ty1 -> [Ty1]

outTy :: ArrowTy Ty1 -> Ty1