Gibbon.L2.Interp

Description

Interpreter reducing L2 programs to values.

Orphan instances

Interp Prog
HACK: we don't have a type-level distinction for when cursors are allowed in the AST. We use L2 as a proxy for this, allowing cursors whenver executing L2, even though this is a bit premature in the compiler pipeline.

Instance details

Methods

interpNoLogs :: RunConfig -> Prog -> String

interpWithStdout :: RunConfig -> Prog -> IO (String, [String])