Gibbon.L2.Syntax

Description

An intermediate language with an effect system that captures traversals.

Extended language L2 with location types.

data E2Ext loc dec

The extension that turns L1 into L2.

Constructors

LetRegionE Region (L (E2 loc dec))
Not used until later on.

LetLocE loc (PreLocExp loc) (L (E2 loc dec))
Bind a new location.

RetE [loc] Var
Return a value together with extra loc values.

FromEndE loc
Bind a location from an EndOf location (for RouteEnds and after)

BoundsCheck Int loc loc
Bytes required, region, write cursor

IndirectionE TyCon DataCon (loc, Var) (loc, Var) (L (E2 loc dec))
An inter-region indirection

type Prog2 = Prog (L Exp2)

Here we only change the types of FUNCTIONS:

type FunDef2 = FunDef (L Exp2)

A function definition with the function's effects.

type FunDefs2 = FunDefs (L Exp2)

type Exp2 = E2 LocVar Ty2

Extended expressions, L2. Monomorphic.

By adding a LocVar decoration, all data constructors, applications, and bindings gain a location annotation.

type E2 l d = PreExp E2Ext l d

L1 expressions extended with L2. This is the polymorphic version. Shorthand for recursions above.

type Ty2 = UrTy LocVar

L1 Types extended with abstract Locations.

data Effect

The side-effect of evaluating a function.

Constructors

Traverse LocVar
The function, during its execution, traverses all of the value living at this location.

data ArrowTy2

Our type for functions grows to include effects, and explicit universal quantification over location/region variables.

Constructors

ArrowTy2

Fields

locVars :: [LRM]

Universally-quantified location params. Only these should be referenced in arrIn/arrOut.

arrIn :: Ty2

Input type for the function.

arrEffs :: Set Effect

These are present-but-empty initially, and the populated by InferEffects.

arrOut :: Ty2

Output type for the function.

locRets :: [LocRet]

L2B feature: multi-valued returns.

data LocRet

Locations (end-witnesses) returned from functions after RouteEnds.

Constructors

EndOf LRM

type LocExp = PreLocExp LocVar

data PreLocExp loc

Define a location in terms of a different location.

Constructors

StartOfLE Region

AfterConstantLE Int loc

AfterVariableLE Var loc

InRegionLE Region

FromEndLE loc

Re-exports

data PreExp (ext :: * -> * -> *) loc dec

The source language. It has pointer-based sums and products, as well as packed algebraic datatypes.

It is parameterized by an a potential extension point.

It is parameterized by loc, the type of locations.

It is parameterized by a decoration, d, attached to every binder.

Constructors

VarE Var
Variable reference

LitE Int
Numeric literal

LitSymE Var
A quoted symbol literal.

AppE Var [loc] (L (PreExp ext loc dec))
Apply a top-level / first-order function. Instantiate its type schema by providing location-variable arguments, if applicable.

PrimAppE (Prim dec) [L (PreExp ext loc dec)]
Primitive applications don't manipulate locations.

LetE (Var, [loc], dec, L (PreExp ext loc dec)) (L (PreExp ext loc dec))
One binding at a time. Allows binding a list of implicit *location* return vales from the RHS, plus a single "real" value. This list of implicit returnsb

IfE (L (PreExp ext loc dec)) (L (PreExp ext loc dec)) (L (PreExp ext loc dec))

MkProdE [L (PreExp ext loc dec)]
Tuple construction

ProjE Int (L (PreExp ext loc dec))
Tuple projection.

CaseE (L (PreExp ext loc dec)) [(DataCon, [(Var, loc)], L (PreExp ext loc dec))]
Case on a datatype. Each bound, unpacked variable lives at a fixed, read-only location.

DataConE loc DataCon [L (PreExp ext loc dec)]
Construct data that may unpack some fields. The location argument, if applicable, is the byte location at which to write the tag for the sum type.

TimeIt (L (PreExp ext loc dec)) dec Bool
The boolean being true indicates this TimeIt is really (iterate _) This iterate form is used for criterion-style benchmarking.

MapE (Var, Ty1, L (PreExp ext loc dec)) (L (PreExp ext loc dec))

FoldE

Fields

initial :: (Var, dec, L (PreExp ext loc dec))

iterator :: (Var, dec, L (PreExp ext loc dec))

body :: L (PreExp ext loc dec)

Ext (ext loc dec)
Extension point for downstream language extensions.

data UrTy a

Types include boxed/pointer-based products as well as unpacked algebraic datatypes. This data is parameterized to allow annotation on Packed types later on.

Constructors

IntTy

BoolTy

ProdTy [UrTy a]
An N-ary tuple

SymDictTy (UrTy a)
A map from SymTy to Ty ^ We allow built-in dictionaries from symbols to a value type.

PackedTy TyCon a
No type arguments to TyCons for now. (No polymorphism.)

ListTy (UrTy a)
These are not fully first class. They are only allowed as the fields of data constructors.

PtrTy
A machine pointer to a complete value in memory. This is decorated with the region it points into, which may affect the memory layout.

CursorTy
A cursor for reading or writing, which may point to an unkwown type or to a fraction of a complete value. It is a machine pointer that can point to any byte.

pattern SymTy :: forall a. UrTy a

Operations on types

progToEnv :: Prog2 -> Env2 Ty2

Abstract some of the differences of top level program types, by having a common way to extract an initial environment. The initial environment has types only for functions.

allLocVars :: ArrowTy2 -> [LocVar]

Retrieve all LocVars from a fn type (Arrow)

inLocVars :: ArrowTy2 -> [LocVar]

outLocVars :: ArrowTy2 -> [LocVar]

outRegVars :: ArrowTy2 -> [LocVar]

inRegVars :: ArrowTy2 -> [LocVar]

substTy :: Map LocVar LocVar -> Ty2 -> Ty2

Apply a variable substitution to a type.

substEffs :: Map LocVar LocVar -> Set Effect -> Set Effect

Apply a substitution to an effect set.

prependArgs :: [UrTy l] -> UrTy l -> UrTy l

Injected cursor args go first in input and output:

stripTyLocs :: UrTy a -> UrTy ()

Remove the extra location annotations.

getTyLocs :: Ty2 -> [LocVar]

Collect all the locations mentioned in a type.

locsInTy :: Ty2 -> [LocVar]

Other helpers

revertToL1 :: Prog2 -> Prog1

occurs :: Var -> L Exp2 -> Bool

Does a variable occurs in an expression ?

mapPacked :: (Var -> l -> UrTy l) -> UrTy l -> UrTy l

depList :: L Exp2 -> [(Var, Var, [Var])]

Build a dependency list which can be later converted to a graph

initFunEnv :: FunDefs2 -> FunEnv Ty2

Orphan instances

Out a => Out (Set a)

Methods

docPrec :: Int -> Set a -> Doc

doc :: Set a -> Doc

docList :: [Set a] -> Doc