Gibbon.Language

Documentation

module Gibbon.Language.Constants

module Gibbon.Language.Syntax

Helpers operating on expressions

mapExt :: (e1 l d -> e2 l d) -> PreExp e1 l d -> PreExp e2 l d

Apply a function to the extension points only.

mapLocs :: (e l2 d -> e l2 d) -> PreExp e l2 d -> PreExp e l2 d

Apply a function to the locations only.

mapExprs :: (e -> e) -> Prog e -> Prog e

Transform the expressions within a program.

mapMExprs :: Monad m => (e -> m e) -> Prog e -> m (Prog e)

Monadic mapExprs.

visitExp :: forall l1 l2 e1 e2 d1 d2. (l1 -> l2) -> (e1 l1 d1 -> e2 l2 d2) -> (d1 -> d2) -> PreExp e1 l1 d1 -> PreExp e2 l2 d2

Apply a function to the locations, extensions, and binder-decorations, respectively.

subst :: HasSubstitutable e l d => Var -> PreExp e l d -> PreExp e l d -> PreExp e l d

Substitute an expression in place of a variable.

substE :: HasSubstitutable e l d => PreExp e l d -> PreExp e l d -> PreExp e l d -> PreExp e l d

Expensive subst that looks for a whole matching sub-EXPRESSION. If the old expression is a variable, this still avoids going under binder.

hasTimeIt :: PreExp e l d -> Bool

Does the expression contain a TimeIt form?

hasSpawns :: PreExp e l d -> Bool

Does the expression contain a SpawnE form?

hasSpawnsProg :: Prog (PreExp e l d) -> Bool

projNonFirst :: (Out l, Out d, Out (e l d)) => Int -> PreExp e l d -> PreExp e l d

Project something which had better not be the first thing in a tuple.

mkProj :: Int -> PreExp e l d -> PreExp e l d

Smart constructor that immediately destroys products if it can: Does NOT avoid single-element tuples.

mkProd :: [PreExp e l d] -> PreExp e l d

Make a product type while avoiding unary products.

mkLets :: [(Var, [loc], dec, PreExp ext loc dec)] -> PreExp ext loc dec -> PreExp ext loc dec

Make a nested series of lets.

flatLets :: [(Var, [l], d, PreExp e l d)] -> PreExp e l d -> PreExp e l d

Alternative version of L1.mkLets that also flattens

tuplizeRefs :: Var -> [Var] -> [d] -> PreExp e l d -> PreExp e l d

Helpers operating on types

mkProdTy :: [UrTy a] -> UrTy a

Same as mkProd, at the type level

projTy :: Out a => Int -> UrTy a -> UrTy a

voidTy :: UrTy a

A makeshift void type.

isProdTy :: UrTy a -> Bool

Are values of this type tuples ?

isNestedProdTy :: UrTy a -> Bool

Do values of this type contain nested tuples ?

isPackedTy :: UrTy a -> Bool

Are values of this type Packed ?

isScalarTy :: UrTy a -> Bool

hasPacked :: Show a => UrTy a -> Bool

Do values of this type contain packed data?

sizeOfTy :: UrTy a -> Maybe Int

Provide a size in bytes, if it is statically known.

primArgsTy :: Prim (UrTy a) -> [UrTy a]

Type of the arguments for a primitive operation.

primRetTy :: Prim (UrTy a) -> UrTy a

Return type for a primitive operation.

tyToDataCon :: Show a => UrTy a -> DataCon

Get the data constructor type from a type, failing if it's not packed

stripTyLocs :: UrTy a -> UrTy ()

isValidListElemTy :: UrTy a -> Bool

Lists of scalars or flat products of scalars are allowed.

getPackedTys :: Show a => UrTy a -> [UrTy a]

Get all packed types in a type.

Misc

assertTriv :: Expression e => e -> a -> a

Ensure that an expression is trivial.

assertTrivs :: Expression e => [e] -> a -> a

List version of assertTriv.

Orphan instances

Renamable Var

Instance details

Methods

gRename :: Map Var Var -> Var -> Var

Renamable a => Renamable (UrTy a)

Instance details

Methods

gRename :: Map Var Var -> UrTy a -> UrTy a

(Out l, Show l, Show d, Out d, Expression (e l d)) => Expression (PreExp e l d)

Instance details

Associated Types

type TyOf (PreExp e l d)

type LocOf (PreExp e l d)

Methods

isTrivial :: PreExp e l d -> Bool

FreeVars (e l d) => FreeVars (PreExp e l d)
Free data variables. Does not include function variables, which currently occupy a different namespace. Does not include location/region variables.

Instance details

Methods

gFreeVars :: PreExp e l d -> Set Var

HasRenamable e l d => Renamable (PreExp e l d)

Instance details

Methods

gRename :: Map Var Var -> PreExp e l d -> PreExp e l d

HasSubstitutable e l d => Substitutable (PreExp e l d)

Instance details

Methods

gSubst :: Var -> PreExp e l d -> PreExp e l d -> PreExp e l d

gSubstE :: PreExp e l d -> PreExp e l d -> PreExp e l d -> PreExp e l d

(Show (), Out (), Expression (e () (UrTy ())), TyOf (e () (UrTy ())) ~ TyOf (PreExp e () (UrTy ())), FunctionTy (UrTy ()), Typeable (e () (UrTy ()))) => Typeable (PreExp e () (UrTy ()))
A Typeable instance for L1 and L3 (L2 defines it's own)

Instance details

Methods

gRecoverType :: DDefs (TyOf (PreExp e () (UrTy ()))) -> Env2 (TyOf (PreExp e () (UrTy ()))) -> PreExp e () (UrTy ()) -> TyOf (PreExp e () (UrTy ()))