Gibbon.Language.Constants

Documentation

redirectionSize :: Int

redirectionTag :: DataCon

redirectionAlt :: Num a => a

indirectionTag :: DataCon

isIndirectionTag :: DataCon -> Bool

indirectionAlt :: Num a => a

toIndrDataCon :: DataCon -> DataCon

fromIndrDataCon :: DataCon -> DataCon

isIndrDataCon :: DataCon -> Bool

mkUnpackerName :: TyCon -> Var

Map a DataCon onto the name of the generated unpack function.

mkPrinterName :: DataCon -> Var

Map a DataCon onto the name of the generated print function.

mkCopyFunName :: DataCon -> Var

isCopyFunName :: Var -> Bool

mkTravFunName :: DataCon -> Var