Gibbon.Language.Syntax

Datatype definitions

type DDefs a = Map Var (DDef a)

type DataCon = String

type TyCon = String

type Tag = Word8

type IsBoxed = Bool

data DDef a

Data type definitions.

Monomorphism: In the extreme case we can strip packed datatypes of all type parameters, or we can allow them to retain type params but require that they always be fully instantiated to monomorphic types in the context of our monomorphic programs.

Here we allow individual to be marked with whether or not they should be boxed. We say that a regular, pointer-based datatype has all-boxed fields, whereas a fully serialized datatype has no boxed fields.

Constructors

DDef

Fields

tyName :: Var

tyArgs :: [TyVar]

dataCons :: [(DataCon, [(IsBoxed, a)])]

lookupDDef :: Out a => DDefs a -> TyCon -> DDef a

Lookup a ddef in its entirety

getConOrdering :: Out a => DDefs a -> TyCon -> [DataCon]

Get the canonical ordering for data constructors, currently based on ordering in the original source code. Takes a TyCon as argument.

getTyOfDataCon :: Out a => DDefs a -> DataCon -> TyCon

Lookup the name of the TyCon that goes with a given DataCon. Must be unique!

lookupDataCon :: Out a => DDefs a -> DataCon -> [a]

Lookup the arguments to a data contstructor.

lkp :: Out a => DDefs a -> DataCon -> (Var, (DataCon, [(IsBoxed, a)]))

Lookup a Datacon. Return (TyCon, (DataCon, [flds]))

lookupDataCon' :: Out a => DDef a -> DataCon -> [a]

Like lookupDataCon but lookup arguments to a data contstructor for a specific instance of a datatype.

lookupDataCon' (Maybe Int) Just = [Int]

insertDD :: DDef a -> DDefs a -> DDefs a

emptyDD :: DDefs a

fromListDD :: [DDef a] -> DDefs a

isVoidDDef :: DDef a -> Bool

Is this an empty type (like 'data Void' in Haskell) ?

Function definitions

class (Out (ArrowTy ty), Show (ArrowTy ty)) => FunctionTy ty where

A type family describing function types.

Minimal complete definition

inTys, outTy

Associated Types

type ArrowTy ty

Methods

inTys :: ArrowTy ty -> [ty]

outTy :: ArrowTy ty -> ty

type FunDefs ex = Map Var (FunDef ex)

A set of top-level recursive function definitions.

data FunDef ex

A function definiton indexed by a type and expression.

Constructors

FunDef

Fields

funName :: Var

funArgs :: [Var]

funTy :: ArrowTy (TyOf ex)

funBody :: ex

insertFD :: FunDef ex -> FunDefs ex -> FunDefs ex

Insert a FunDef into FunDefs. Raise an error if a function with the same name already exists.

fromListFD :: [FunDef ex] -> FunDefs ex

initFunEnv :: FunDefs a -> TyEnv (ArrowTy (TyOf a))

Programs

data Prog ex

Complete programs include datatype definitions:

For evaluating a complete program, main's type will be an Int or a datatype. For running a pass benchmark, main will be Nothing and we will expect a "benchmark" function definition which consumes an appropriate packed AST datatype.

Constructors

Prog

Fields

ddefs :: DDefs (TyOf ex)

fundefs :: FunDefs ex

mainExp :: Maybe (ex, TyOf ex)

progToEnv :: Prog a -> Env2 (TyOf a)

Abstract some of the differences of top level program types, by having a common way to extract an initial environment. The initial environment has types only for functions.

getFunTy :: Var -> Prog ex -> ArrowTy (TyOf ex)

Look up the input/output type of a top-level function binding.

Generic operations

class FreeVars a where

Expression and program types which support a notion of free variables.

Minimal complete definition

gFreeVars

Methods

gFreeVars :: a -> Set Var

Return a set of free TERM variables. Does not return location variables.

class (Show e, Out e, FreeVars e) => Expression e where

A generic interface to expressions found in different phases of the compiler.

Minimal complete definition

isTrivial

Associated Types

type TyOf e

The type representation used in this expression.

type LocOf e

The location (variable) representation used in this expression.

Methods

isTrivial :: e -> Bool

Is an expression considered trivial (duplicatable by the compiler)?

type Binds e = (Var, [LocOf e], TyOf e, e)

class Expression e => Flattenable e where

IRs amenable to flattening

Minimal complete definition

gFlattenExp, gFlattenGatherBinds

Methods

gFlattenExp :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> PassM e

Process an expression into a fully-flattened expression which typically includes a larger number of temporary, local variable bindings.

gFlattenGatherBinds :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> PassM ([Binds e], e)

A private method. Gather the bindings from a subexpression, but do not "discharge" them by creating a let expression. They are in order, so later may depend on earlier.

class Expression e => Simplifiable e where

IRs amenable to simplification/inlineTrivs

Minimal complete definition

gInlineTrivExp

Methods

gInlineTrivExp :: DDefs (TyOf e) -> e -> e

class Expression e => Typeable e where

This is NOT a replacement for any typechecker. This is supposed to be used just to recover the type of an expression in a type-environment Without this, we cannot have truly generic implementation of the Flattenable class, since we need to know the type of an expression before we discharge it with a LetE

Minimal complete definition

gRecoverType

Methods

gRecoverType :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> TyOf e

Environments

type TyEnv a = Map Var a

A simple type environment

data Env2 a

A common currency for a two part environment consisting of function bindings and regular value bindings.

Constructors

Env2

Fields

vEnv :: TyEnv a

fEnv :: TyEnv (ArrowTy a)

emptyEnv2 :: Env2 a

extendVEnv :: Var -> a -> Env2 a -> Env2 a

Extend non-function value environment.

extendsVEnv :: Map Var a -> Env2 a -> Env2 a

Extend multiple times in one go.

lookupVEnv :: Out a => Var -> Env2 a -> a

extendFEnv :: Var -> ArrowTy a -> Env2 a -> Env2 a

Extend function type environment.

lookupFEnv :: Out (ArrowTy a) => Var -> Env2 a -> ArrowTy a

Expresssions and thier types

data PreExp (ext :: * -> * -> *) loc dec

The source language. It has pointer-based sums and products, as well as packed algebraic datatypes.

It is parameterized by an a potential extension point.

It is parameterized by loc, the type of locations.

It is parameterized by a decoration, d, attached to every binder.

Constructors

VarE Var
Variable reference

LitE Int
Numeric literal

LitSymE Var
A quoted symbol literal.

AppE Var [loc] [L (PreExp ext loc dec)]
Apply a top-level / first-order function. Instantiate its type schema by providing location-variable arguments, if applicable.

PrimAppE (Prim dec) [L (PreExp ext loc dec)]
Primitive applications don't manipulate locations.

LetE (Var, [loc], dec, L (PreExp ext loc dec)) (L (PreExp ext loc dec))
One binding at a time. Allows binding a list of implicit *location* return vales from the RHS, plus a single "real" value. This list of implicit returnsb

IfE (L (PreExp ext loc dec)) (L (PreExp ext loc dec)) (L (PreExp ext loc dec))

MkProdE [L (PreExp ext loc dec)]
Tuple construction

ProjE Int (L (PreExp ext loc dec))
Tuple projection.

CaseE (L (PreExp ext loc dec)) [(DataCon, [(Var, loc)], L (PreExp ext loc dec))]
Case on a datatype. Each bound, unpacked variable lives at a fixed, read-only location.

DataConE loc DataCon [L (PreExp ext loc dec)]
Construct data that may unpack some fields. The location argument, if applicable, is the byte location at which to write the tag for the sum type.

TimeIt (L (PreExp ext loc dec)) dec Bool
The boolean being true indicates this TimeIt is really (iterate _) This iterate form is used for criterion-style benchmarking.

ParE (L (PreExp ext loc dec)) (L (PreExp ext loc dec))
Parallel tuple combitor.

MapE (Var, dec, L (PreExp ext loc dec)) (L (PreExp ext loc dec))

FoldE

Fields

initial :: (Var, dec, L (PreExp ext loc dec))

iterator :: (Var, dec, L (PreExp ext loc dec))

body :: L (PreExp ext loc dec)

Ext (ext loc dec)
Extension point for downstream language extensions.

data Prim ty

Some of these primitives are (temporarily) tagged directly with their return types.

Constructors

AddP

SubP

MulP
May need more numeric primitives...

DivP

ModP
Integer division and modulus

ExpP
Exponentiation

RandP
Generate a random number. Translates to 'rand()' in C. TEMP: It's a side-effect, and should be removed.

EqSymP
Equality on Sym

EqIntP
Equality on Int

LtP

GtP
(and () for Int's

LtEqP

GtEqP
and=

OrP

AndP

SymAppend
A quick hack till we have deterministic gensym

DictInsertP ty
takes dict, k,v; annotated with element type

DictLookupP ty
takes dict,k errors if absent; annotated with element type

DictEmptyP ty
annotated with element type to avoid ambiguity

DictHasKeyP ty
takes dict,k; returns a Bool, annotated with element type

ErrorP String ty
crash and issue a static error message. To avoid needing inference, this is labeled with a return type.

SizeParam

MkTrue
Zero argument constructor.

MkFalse
Zero argument constructor.

ReadPackedFile (Maybe FilePath) TyCon (Maybe Var) ty
Read (mmap) a binary file containing packed data. This must be annotated with the type of the file being read. The Ty tracks the type as the program evolvels (first PackedTy then CursorTy). The TyCon tracks the original type name. The variable represents the region that this file will be mapped to, and is set by InferLocations.

PEndOf
Returns the end-of cursor for some packed value. These pointers serve as random access nodes to the value that is packed after this one.

data UrTy a

Types include boxed/pointer-based products as well as unpacked algebraic datatypes. This data is parameterized to allow annotation on Packed types later on.

Constructors

IntTy

BoolTy

ProdTy [UrTy a]
An N-ary tuple

SymDictTy (UrTy a)
A map from SymTy to Ty ^ We allow built-in dictionaries from symbols to a value type.

PackedTy TyCon a
No type arguments to TyCons for now. (No polymorphism.)

ListTy (UrTy a)
These are not fully first class. They are only allowed as the fields of data constructors.

PtrTy
A machine pointer to a complete value in memory. This is decorated with the region it points into, which may affect the memory layout.

CursorTy
A cursor for reading or writing, which may point to an unkwown type or to a fraction of a complete value. It is a machine pointer that can point to any byte.

pattern SymTy :: forall a. UrTy a

Helpers operating on expressions

mapExt :: (e1 l d -> e2 l d) -> PreExp e1 l d -> PreExp e2 l d

Apply a function to the extension points only.

mapLocs :: (e l2 d -> e l2 d) -> PreExp e l2 d -> PreExp e l2 d

Apply a function to the locations only.

mapExprs :: (e -> e) -> Prog e -> Prog e

Transform the expressions within a program.

mapMExprs :: Monad m => (e -> m e) -> Prog e -> m (Prog e)

Monadic mapExprs.

visitExp :: forall l1 l2 e1 e2 d1 d2. (l1 -> l2) -> (e1 l1 d1 -> e2 l2 d2) -> (d1 -> d2) -> PreExp e1 l1 d1 -> PreExp e2 l2 d2

Apply a function to the locations, extensions, and binder-decorations, respectively.

subst :: (Eq d, Eq l, Eq (e l d)) => Var -> L (PreExp e l d) -> L (PreExp e l d) -> L (PreExp e l d)

Substitute an expression in place of a variable.

substE :: (Eq d, Eq l, Eq (e l d)) => L (PreExp e l d) -> L (PreExp e l d) -> L (PreExp e l d) -> L (PreExp e l d)

Expensive subst that looks for a whole matching sub-EXPRESSION. If the old expression is a variable, this still avoids going under binder.

hasTimeIt :: L (PreExp e l d) -> Bool

Does the expression contain a TimeIt form?

projNonFirst :: (Out l, Out d, Out (e l d)) => Int -> L (PreExp e l d) -> L (PreExp e l d)

Project something which had better not be the first thing in a tuple.

mkProj :: Int -> L (PreExp e l d) -> L (PreExp e l d)

Smart constructor that immediately destroys products if it can: Does NOT avoid single-element tuples.

mkProd :: [L (PreExp e l d)] -> L (PreExp e l d)

Make a product type while avoiding unary products.

mkLets :: [(Var, [loc], dec, L (PreExp ext loc dec))] -> L (PreExp ext loc dec) -> L (PreExp ext loc dec)

Make a nested series of lets.

flatLets :: [(Var, [l], d, L (PreExp e l d))] -> L (PreExp e l d) -> L (PreExp e l d)

Alternative version of L1.mkLets that also flattens

tuplizeRefs :: Var -> [Var] -> [d] -> L (PreExp e l d) -> L (PreExp e l d)

Helpers operating on types

mkProdTy :: [UrTy a] -> UrTy a

Same as mkProd, at the type level

projTy :: Out a => Int -> UrTy a -> UrTy a

voidTy :: UrTy a

A makeshift void type.

isProdTy :: UrTy a -> Bool

Are values of this type tuples ?

isNestedProdTy :: UrTy a -> Bool

Do values of this type contain nested tuples ?

isPackedTy :: UrTy a -> Bool

Are values of this type Packed ?

hasPacked :: Show a => UrTy a -> Bool

Do values of this type contain packed data?

sizeOfTy :: UrTy a -> Maybe Int

Provide a size in bytes, if it is statically known.

primArgsTy :: Prim (UrTy a) -> [UrTy a]

Type of the arguments for a primitive operation.

primRetTy :: Prim (UrTy a) -> UrTy a

Return type for a primitive operation.

dummyCursorTy :: UrTy a

tyToDataCon :: Show a => UrTy a -> DataCon

Get the data constructor type from a type, failing if it's not packed

Misc

assertTriv :: Expression e => L e -> a -> a

Ensure that an expression is trivial.

assertTrivs :: Expression e => [L e] -> a -> a

List version of assertTriv.