Gibbon.NewL2.Syntax

Description

An intermediate language with an effect system that captures traversals.

Extended language L2 with location types.

data E2Ext loc dec

The extension that turns L1 into L2.

Constructors

LetRegionE Region RegionSize (Maybe RegionType) (E2 loc dec)
Allocate a new region.

LetParRegionE Region RegionSize (Maybe RegionType) (E2 loc dec)
Allocate a new region for parallel allocations.

LetLocE LocVar (PreLocExp loc) (E2 loc dec)
Bind a new location.

RetE [loc] Var
Return a value together with extra loc values.

FromEndE loc
Bind a location from an EndOf location (for RouteEnds and after).

BoundsCheck Int loc loc

AddFixed Var Int

IndirectionE TyCon DataCon (loc, loc) (loc, loc) (E2 loc dec)
A indirection node.

StartOfPkdCursor Var

TagCursor Var Var

GetCilkWorkerNum
Translates to __cilkrts_get_worker_number().

LetAvail [Var] (E2 loc dec)
These variables are available to use before the join point.

AllocateTagHere LocVar TyCon

AllocateScalarsHere LocVar
A marker which tells subsequent a compiler pass where to move the tag and scalar field allocations so that they happen before any of the subsequent packed fields.

SSPush SSModality LocVar LocVar TyCon

SSPop SSModality LocVar LocVar
Spill and restore from the shadow-stack.

type Prog2 = Prog Exp2

type DDefs2 = DDefs Ty2

type DDef2 = DDef Ty2

type FunDef2 = FunDef Exp2

type FunDefs2 = FunDefs Exp2

type Exp2 = PreExp E2Ext LocArg Ty2

Extended expressions, L2.

By adding a LocVar decoration, all data constructors, applications, and bindings gain a location annotation.

newtype Ty2

L1 Types extended with abstract Locations.

Constructors

MkTy2

Fields

unTy2 :: UrTy LocVar

data Effect

The side-effect of evaluating a function.

Constructors

Traverse LocVar
The function, during its execution, traverses all of the value living at this location.

data ArrowTy2 ty2

Our type for functions grows to include effects, and explicit universal quantification over location/region variables.

Constructors

ArrowTy2

Fields

locVars :: [LRM]

Universally-quantified location params. Only these should be referenced in arrIn/arrOut.

arrIns :: [ty2]

Input type for the function.

arrEffs :: Set Effect

These are present-but-empty initially, and the populated by InferEffects.

arrOut :: ty2

Output type for the function.

locRets :: [LocRet]

L2B feature: multi-valued returns.

hasParallelism :: Bool

Does this function have parallelism

data LocRet

Locations (end-witnesses) returned from functions after RouteEnds.

Constructors

EndOf LRM

data LocArg

Constructors

Loc LREM

EndWitness LREM Var

Reg RegVar Modality

EndOfReg RegVar Modality RegVar

EndOfReg_Tagged RegVar

type LocExp = PreLocExp LocArg

data PreLocExp loc

Define a location in terms of a different location.

Constructors

StartOfRegionLE Region

AfterConstantLE Int loc

AfterVariableLE Var loc Bool

InRegionLE Region

FreeLE

FromEndLE loc

Regions and locations

type LocVar = Var

Abstract location variables.

data Region

An abstract region identifier. This is used inside type signatures and elsewhere.

Constructors

GlobR Var Multiplicity
A global region with lifetime equal to the whole program.

DynR Var Multiplicity
A dynamic region that may be created or destroyed, tagged by an identifier.

VarR Var
A region metavariable that can range over either global or dynamic regions.

MMapR Var
A region that doesn't result in an (explicit) memory allocation. It merely ensures that there are no free locations in the program.

data Modality

The modality of locations and cursors: input/output, for reading and writing, respectively.

Constructors

Input

Output

data LRM

A location and region, together with modality.

Constructors

LRM

Fields

lrmLoc :: LocVar

lrmReg :: Region

lrmMode :: Modality

data LREM

Constructors

LREM

Fields

lremLoc :: LocVar

lremReg :: RegVar

lremEndReg :: RegVar

lremMode :: Modality

data Multiplicity

Region variants (multiplicities)

Constructors

Bounded Int
Contain a finite number of values and can be stack-allocated.

Infinite
Consist of a linked list of buffers, spread throughout memory (though possible constrained to 4GB regions). Writing into these regions requires bounds-checking. The buffers can start very small at the head of the list, but probably grow geometrically in size, making the cost of traversing all of them logarithmic.

BigInfinite
These regions are infinite, but also have the expectation of containing many values. Thus we give them large initial page sizes. This is also could be the appropriate place to use mmap to grow the region and to establish guard places.

data RegionSize

Constructors

BoundedSize Int

Undefined

data RegionType

Constructors

IndirectionFree

RightwardLocalIndirections

LocalIndirections

NoSharing

regionToVar :: Region -> Var

Operations on types

allLocVars :: ArrowTy2 ty2 -> [LocVar]

Retrieve all LocVars from a fn type (Arrow)

inLocVars :: ArrowTy2 ty2 -> [LocVar]

outLocVars :: ArrowTy2 ty2 -> [LocVar]

outRegVars :: ArrowTy2 ty2 -> [LocVar]

inRegVars :: ArrowTy2 ty2 -> [LocVar]

allRegVars :: ArrowTy2 ty2 -> [LocVar]

substLoc :: Map LocVar LocVar -> Ty2 -> Ty2

Apply a location substitution to a type.

substLocs :: Map LocVar LocVar -> [Ty2] -> [Ty2]

List version of substLoc.

substEff :: Map LocVar LocVar -> Effect -> Effect

Apply a substitution to an effect.

substEffs :: Map LocVar LocVar -> Set Effect -> Set Effect

Apply a substitution to an effect set.

extendPatternMatchEnv :: HasCallStack => DataCon -> DDefs Ty2 -> [Var] -> [LocVar] -> Env2 Ty2 -> Env2 Ty2

Extend an environment for a pattern match. E.g.

data Foo = MkFoo Int Foo | ...

case foo1 of MkFoo (i:loc1) (f:loc2) -> new_env2 = extendPatternMatchEnv [loc1,loc2] old_env2

locsInTy :: Ty2 -> [LocVar]

Collect all the locations mentioned in a type.

dummyTyLocs :: Applicative f => UrTy () -> f (UrTy LocVar)

allFreeVars :: Exp2 -> Set Var

freeLocVars :: Exp2 -> [Var]

toLocVar :: LocArg -> LocVar

fromLRM :: LRM -> LREM

Other helpers

revertToL1 :: Prog2 -> Prog1

occurs :: Set Var -> Exp2 -> Bool

Does a variable occur in an expression ?

N.B. it only looks for actual variables, not LocVar's or RegionVar's.

mapPacked :: (Var -> l -> UrTy l) -> UrTy l -> UrTy l

constPacked :: UrTy a1 -> UrTy a2 -> UrTy a1

depList :: Exp2 -> [(Var, Var, [Var])]

Build a dependency list which can be later converted to a graph

changeAppToSpawn :: (Eq loc, Eq dec) => Var -> [PreExp E2Ext loc dec] -> PreExp E2Ext loc dec -> PreExp E2Ext loc dec

toEndFromTaggedV :: Var -> Var

toTagV :: Var -> Var

module Gibbon.Language