Gibbon.Passes.AddTraversals

Documentation

addTraversals :: Prog2 -> PassM Prog2

needsTraversal :: DDefs Ty2 -> FunDefs2 -> Env2 Ty2 -> (DataCon, [(Var, LocVar)], L Exp2) -> Maybe [(Var, LocVar)]

Collect all non-static items that need to be traversed (uses InferEffects).

If we cannot unpack all the pattern matched variables: (1) Everything after the first packed element should be unused in the RHS (2) Otherwise, we must traverse the first (n-1) packed elements