Gibbon.Passes.Codegen

Description

The final pass of the compiler: generate C code.

Documentation

codegenProg :: Config -> Prog -> IO String

Compile a program to C code which has the side effect of the "main" expression in that program.

The boolean flag is true when we are compiling in Packed mode.

harvestStructTys :: Prog -> Set [Ty]

Harvest all struct tys. All product types used anywhere in the program.

makeName :: [Ty] -> String

rewriteReturns :: Tail -> [(Var, Ty)] -> Tail

Replace returns with assignments to a given set of destinations.