Gibbon.Passes.Flatten

Description

Put the program in A-normal form where only varrefs and literals are allowed in operand position.

Documentation

flattenL0 :: Prog0 -> PassM Prog0

flattenL1 :: Prog1 -> PassM Prog1

Flatten ensures that function operands are "trivial".

In the process, it also lifts lets out of case scrutinees, if conditions, and tuple operands.

Note that it does not require tail expressions to be trivial. For example, it allows AppE and PrimAppE in the body of a let-expression.

flattenL2 :: Flattenable (E2Ext Var (UrTy LocVar)) => Prog2 -> PassM Prog2

flattenL3 :: Prog3 -> PassM Prog3

Orphan instances

FlattenDeps e l d => Flattenable (PreExp e l d)

Instance details

Methods

gFlattenExp :: DDefs (TyOf (PreExp e l d)) -> Env2 (TyOf (PreExp e l d)) -> PreExp e l d -> PassM (PreExp e l d)

gFlattenGatherBinds :: DDefs (TyOf (PreExp e l d)) -> Env2 (TyOf (PreExp e l d)) -> PreExp e l d -> PassM ([Binds (PreExp e l d)], PreExp e l d)