Gibbon.Passes.InferEffects

Description

An intermediate language with an effect system that captures traversals.

ASSUMES that the flatten pass has run, and thus we have trivial AppE operands.

Documentation

inferEffects :: Prog2 -> PassM Prog2

inferExp :: DDefs Ty2 -> FunEnv2 -> TyEnv Ty2 -> L Exp2 -> (Set Effect, Maybe LocVar)