Gibbon.Passes.InferMultiplicity

Description

Infer region multiplicities

During inference, regions are merely annotated with a region metavariable. InferMultiplicity takes the next step and decides the region scope (global/dynamic) and also assigns it a multiplicity.

Documentation

inferRegScope :: Multiplicity -> Prog2 -> SyM Prog2

Infer multiplicity for a program annotated with regions & locations

inferRegScopeExp :: Multiplicity -> L Exp2 -> L Exp2

Decide if a region should be global or local (dynamic).

Dynamic regions are stack allocated and automatically freed