Gibbon.Passes.InferRegionScope

Description

Infer region multiplicities

During inference, regions are merely annotated with a region metavariable. InferRegionScope takes the next step and decides the region scope (global/dynamic) and also assigns it a multiplicity.

Documentation

inferRegScope :: Prog2 -> PassM Prog2

Infer region scope and multiplicity for a program annotated with regions and locations.

inferRegScopeExp :: Map Region Region -> Exp2 -> PassM Exp2

Decide if a region should be global or local (dynamic).

Dynamic regions are stack allocated and automatically freed