Gibbon.Passes.Lower

Description

Lowering L3 to the target language.

Documentation

lower :: Prog3 -> PassM Prog

Convert into the target language. This does not make much of a change, but it checks the changes that have already occurred.

The only substantitive conversion here is of tupled arguments to multiple argument functions.

First argument indicates (1) whether we're inpacked mode, and (2) the pre-cursorize type of the mainExp, if there is a mainExp.

getTagOfDataCon :: Out a => DDefs a -> DataCon -> Tag

Look up the numeric tag for a dataCon