Gibbon.Passes.ReorderScalarWrites

Documentation

reorderScalarWrites :: Prog3 -> PassM Prog3

writeOrderMarkers :: Prog2 -> PassM Prog2

Inserts markers which tells subsequent a compiler pass where to move the tag and scalar field allocations so that they happen before any of the subsequent packed fields.