Gibbon.Passes.Unariser

Documentation

unariser :: Prog3 -> PassM Prog3

This pass gets ready for Lower by converting most uses of projection and tuple-construction into finer-grained bindings.

OUTPUT INVARIANTS:

only flat tuples as function arguments (no nesting), all arguments immediately present, e.g. `AppE "f" (MkProd [x,y,z])` rather than `AppE "f" (MkProdE [x,MkProdE[y,z]])`

The only MkProdE allowed outside of function operands is within return/tail position (of a function or If branch).

Primitives are allowed to return tuples, but are let-bound (these will turn into LetPrimCall). The references to these tuples are all of the form `ProjE i (VarE v)` and they are then transformed to varrefs in lower.

unariserExp :: DDefs Ty3 -> ProjStack -> Env2 Ty3 -> L Exp3 -> PassM (L Exp3)