Gibbon.Common

Description

Utilities and common types.

Global constants or conventions

mkUnpackerName :: TyCon -> Var

Map a DataCon onto the name of the generated unpack function.

mkPrinterName :: DataCon -> Var

Map a DataCon onto the name of the generated print function.

Type and Data constructors

type DataCon = String

type TyCon = String

type IsBoxed = Bool

Variables and gensyms

newtype Var

Constructors

Var Symbol

fromVar :: Var -> String

toVar :: String -> Var

varAppend :: Var -> Var -> Var

String concatenation on variables.

data SyM a

gensym :: Var -> SyM Var

Generate a unique symbol by attaching a numeric suffix.

genLetter :: SyM Var

Generate alphabetic variables a,b,...

runSyM :: Int -> SyM a -> (a, Int)

cleanFunName :: Var -> Var

Filter out non-C compatible characters. This naively assumes it will get no conflicts. Which may be correct if function names were gensym'd also....

mkCopyFunName :: DataCon -> Var

isCopyFunName :: Var -> Bool

Regions and locations

type LocVar = Var

Abstract location variables.

data Region

An abstract region identifier. This is used inside type signatures and elsewhere.

Constructors

GlobR Var Multiplicity
A global region with lifetime equal to the whole program.

DynR Var Multiplicity
A dynamic region that may be created or destroyed, tagged by an identifier.

VarR Var
A region metavariable that can range over either global or dynamic regions.

data Modality

The modality of locations and cursors: input/output, for reading and writing, respectively.

Constructors

Input

Output

data LRM

A location and region, together with modality.

Constructors

LRM

Fields

lrmLoc :: LocVar

lrmReg :: Region

lrmMode :: Modality

dummyLRM :: LRM

A designated doesn't-really-exist-anywhere location.

data Multiplicity

Region variants (multiplicities)

Constructors

Bounded
Contain a finite number of values and can be stack-allocated.

Infinite
Consist of a linked list of buffers, spread throughout memory (though possible constrained to 4GB regions). Writing into these regions requires bounds-checking. The buffers can start very small at the head of the list, but probably grow geometrically in size, making the cost of traversing all of them logarithmic.

BigInfinite
These regions are infinite, but also have the expectation of containing many values. Thus we give them large initial page sizes. This is also could be the appropriate place to use mmap to grow the region and to establish guard places.

regionVar :: Region -> Var

Environment and helpers

data Env2 a

A common currency for a two part environment consisting of function bindings and regular value bindings.

Constructors

Env2 (Map Var a) (FunEnv a)

type FunEnv a = Map Var (a, a)

Type environment for function defs only. This works with type representations that do not include arrow types.

vEnv :: Env2 a -> Map Var a

fEnv :: Env2 a -> FunEnv a

extendVEnv :: Var -> a -> Env2 a -> Env2 a

Extend non-function value environment.

extendsVEnv :: Map Var a -> Env2 a -> Env2 a

Extend multiple times in one go.

extendFEnv :: Var -> (a, a) -> Env2 a -> Env2 a

Extend function type environment.

emptyEnv2 :: Env2 a

lookupVEnv :: Out a => Var -> Env2 a -> a

Interpreter configuration

data RunConfig

Runtime configuration for executing interpreters.

Constructors

RunConfig

Fields

rcSize :: Int

rcIters :: Word64

rcDbg :: Int

rcCursors :: Bool

Do we support cursors in L1.Exp at this point in the compiler.

getRunConfig :: [String] -> IO RunConfig

We currently use the hacky approach of using env vars OR command line args to set the two universal benchmark params: SIZE and ITERS.

This takes extra, optional command line args [Size, Iters] provided after the file to process on the command line. If these are not present it

data Store

A store is an address space full of buffers.

Constructors

Store (IntMap Buffer)

data Buffer

Constructors

Buffer (Seq SerializedVal)

data SerializedVal

Constructors

SerTag Word8 DataCon

SerInt Int

data Value

It's a first order language with simple values.

Constructors

VInt Int

VBool Bool

VDict (Map Value Value)

VProd [Value]

VPacked DataCon [Value]

VCursor
Cursor are a pointer into the Store plus an offset into the Buffer.

Fields

bufID :: Int

offset :: Int

type Log = Builder

Data definitions

data DDef a

Data type definitions.

Monomorphism: In the extreme case we can strip packed datatypes of all type parameters, or we can allow them to retain type params but require that they always be fully instantiated to monomorphic types in the context of our monomorphic programs.

Here we allow individual to be marked with whether or not they should be boxed. We say that a regular, pointer-based datatype has all-boxed fields, whereas a fully serialized datatype has no boxed fields.

Constructors

DDef

Fields

tyName :: Var

dataCons :: [(DataCon, [(IsBoxed, a)])]

type DDefs a = Map Var (DDef a)

fromListDD :: [DDef a] -> DDefs a

emptyDD :: DDefs a

insertDD :: DDef a -> DDefs a -> DDefs a

lookupDDef :: Out a => DDefs a -> Var -> DDef a

Lookup a ddef in its entirety

lookupDataCon :: Out a => DDefs a -> DataCon -> [a]

Lookup the arguments to a data contstructor.

getConOrdering :: Out a => DDefs a -> TyCon -> [DataCon]

Get the canonical ordering for data constructors, currently based on ordering in the original source code. Takes a TyCon as argument.

getTyOfDataCon :: Out a => DDefs a -> DataCon -> TyCon

Lookup the name of the TyCon that goes with a given DataCon. Must be unique!

getTagOfDataCon :: Out a => DDefs a -> DataCon -> Word8

Look up the numeric tag for a dataCon

type Tag = Word8

Redirections and indirections

redirectionSize :: Int

redirectionTag :: DataCon

redirectionAlt :: Num a => a

toIndrDataCon :: DataCon -> DataCon

fromIndrDataCon :: DataCon -> DataCon

isIndrDataCon :: DataCon -> Bool

indirectionTag :: DataCon

indirectionAlt :: Num a => a

isIndirectionTag :: DataCon -> Bool

Misc helpers

(#) :: (Ord a, Out a, Out b, Show a) => Map a b -> a -> b

(!!!) :: Out a => [a] -> Int -> a

fragileZip :: (Show a, Show b) => [a] -> [b] -> [(a, b)]

fragileZip' :: (Show a, Show b) => [a] -> [b] -> String -> [(a, b)]

Like fragileZip, but takes a custom error message.

sdoc :: Out a => a -> String

Handy combination of show and doc

ndoc :: Out a => a -> String

Like sdoc but inserts newline if it is longish.

abbrv :: Out a => Int -> a -> String

Like ndoc/sdoc but cut it off with "..." after a char limit.

l :: a -> L a

Debugging/logging:

dbgLvl :: Int

Debugging flag shared by all modules. This is activated by setting the environment variable DEBUG=1..5

dbgPrint :: Int -> String -> IO ()

Print if the debug level is at or above a threshold.

dbgPrintLn :: Int -> String -> IO ()

dbgTrace :: Int -> String -> a -> a

Conditional version of Debug.Trace.trace

dbgTraceIt :: Show a => Int -> String -> a -> a

minChatLvl :: Int

We should not create chatter below this level. DEBUG=1 is used for assertions only, not chatter.

internalError :: String -> a

An error that is OUR FAULT, i.e. an internal bug in the compiler.

Establish conventions for the output of #lang gibbon:

truePrinted :: String

For now this is designed to match the Racket output of "#lang gibbon" which itself should change once we implement a custom printer.

falsePrinted :: String

Orphan instances

Generic Loc

Instance details

Associated Types

type Rep Loc :: * -> *

Methods

from :: Loc -> Rep Loc x

to :: Rep Loc x -> Loc

Generic Pos

Instance details

Associated Types

type Rep Pos :: * -> *

Methods

from :: Pos -> Rep Pos x

to :: Rep Pos x -> Pos

Out Word8

Instance details

Methods

docPrec :: Int -> Word8 -> Doc

doc :: Word8 -> Doc

docList :: [Word8] -> Doc

Out Loc

Instance details

Methods

docPrec :: Int -> Loc -> Doc

doc :: Loc -> Doc

docList :: [Loc] -> Doc

Out Pos

Instance details

Methods

docPrec :: Int -> Pos -> Doc

doc :: Pos -> Doc

docList :: [Pos] -> Doc

Read t => Read (L t)
Orphaned instance: read without source locations.

Instance details

Methods

readsPrec :: Int -> ReadS (L t)

readList :: ReadS [L t]

readPrec :: ReadPrec (L t)

readListPrec :: ReadPrec [L t]

Out a => Out (IntMap a)

Instance details

Methods

docPrec :: Int -> IntMap a -> Doc

doc :: IntMap a -> Doc

docList :: [IntMap a] -> Doc

Out a => Out (Seq a)

Instance details

Methods

docPrec :: Int -> Seq a -> Doc

doc :: Seq a -> Doc

docList :: [Seq a] -> Doc

(Out k, Out v) => Out (Map k v)

Instance details

Methods

docPrec :: Int -> Map k v -> Doc

doc :: Map k v -> Doc

docList :: [Map k v] -> Doc