Gibbon.Common

Description

Utilities and common types.

Variables

newtype Var

Constructors

Var Symbol

type LocVar = Var

Abstract location variables.

type RegVar = Var

Abstract region variables.

fromVar :: Var -> String

toVar :: String -> Var

varAppend :: Var -> Var -> Var

String concatenation on variables.

toEndV :: Var -> Var

toSeqV :: Var -> Var

cleanFunName :: Var -> Var

Filter out non-C compatible characters. This naively assumes it will get no conflicts. Which may be correct if function names were gensym'd also....

data TyVar

Type variables that enable polymorphism.

Constructors

BoundTv Var

SkolemTv String Int

UserTv Var

isUserTv :: TyVar -> Bool

data Symbol

intern :: String -> Symbol

unintern :: Symbol -> String

Gensym monad

data SyM a

gensym :: MonadState Int m => Var -> m Var

Generate a unique symbol by attaching a numeric suffix.

gensym_tag :: MonadState Int m => Var -> String -> m Var

genLetter :: MonadState Int m => m Var

An infinite alphabet generator: a,b, ... ,z,a0, ...

newUniq :: MonadState Int m => m Int

A fresh int.

runSyM :: Int -> SyM a -> (a, Int)

PassM monad

data PassM a

The monad used by core Gibbon passes to access Config and other shared state.

runPassM :: Config -> Int -> PassM a -> (a, Int)

defaultRunPassM :: PassM a -> (a, Int)

A convenient wrapper over runPassM.

defaultPackedRunPassM :: PassM a -> (a, Int)

A convenient wrapper over runPassM for running passes in packed mode.

getDynFlags :: MonadReader Config m => m DynFlags

Gibbon configuration

data Config

Overall configuration of the compiler, as determined by command line arguments and possible environment variables.

Constructors

Config

Fields

input :: Input

mode :: Mode

How to run, which backend.

benchInput :: Maybe FilePath

What packed, binary .gpkd file to use as input.

arrayInput :: Maybe FilePath

What array file to use as input.

verbosity :: Int

Debugging output, equivalent to DEBUG env var.

cc :: String

C compiler to use

optc :: String

Options to the C compiler

cfile :: Maybe FilePath

Optional override to destination .c file.

exefile :: Maybe FilePath

Optional override to destination binary file.

backend :: Backend

Compilation backend used

dynflags :: DynFlags

srcFile :: Maybe FilePath

The file being compiled by Gibbon.

data Input

What input format to expect on disk.

Constructors

Haskell

SExpr

Unspecified

data Mode

How far to run the compiler/interpreter.

Constructors

ToParse
Parse and then stop

ToC
Compile to C

ToExe
Compile to C then build a binary.

RunExe
Compile to executable then run.

Interp2
Interp late in the compiler pipeline.

Interp1
Interp early.

ToMPL
Compile to SML (mlton dialect)

ToMPLExe
Compile to SML & compile with MPL

RunMPL
Compile to SML & compile with MPL & run

Bench Var
Benchmark a particular function applied to the packed data within an input file.

BenchInput FilePath
Hardcode the input file to the benchmark in the C code.

data Backend

Compilation backend used

Constructors

C

LLVM

defaultConfig :: Config

data RunConfig

Runtime configuration for executing interpreters.

Constructors

RunConfig

Fields

rcSize :: Int

rcIters :: Word64

rcDbg :: Int

rcCursors :: Bool

Do we support cursors in L1.Exp at this point in the compiler.

getRunConfig :: [String] -> IO RunConfig

We currently use the hacky approach of using env vars OR command line args to set the two universal benchmark params: SIZE and ITERS.

This takes extra, optional command line args [Size, Iters] provided after the file to process on the command line. If these are not present it

defaultRunConfig :: RunConfig

getGibbonConfig :: MonadReader Config m => m Config

Misc helpers

data SSModality

Constructors

Read

Write

(#) :: (Ord a, Out a, Out b, Show a, HasCallStack) => Map a b -> a -> b

(!!!) :: (Out a, HasCallStack) => [a] -> Int -> a

fragileZip :: (Show a, Show b, HasCallStack) => [a] -> [b] -> [(a, b)]

fragileZip' :: (Show a, Show b, HasCallStack) => [a] -> [b] -> String -> [(a, b)]

Like fragileZip, but takes a custom error message.

sdoc :: Out a => a -> String

Handy combination of show and doc

ndoc :: Out a => a -> String

Like sdoc but inserts newline if it is longish.

abbrv :: Out a => Int -> a -> String

Like ndoc/sdoc but cut it off with "..." after a char limit.

lookup3 :: (Eq k, Show k, Show a, Show b) => k -> [(k, a, b)] -> (k, a, b)

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

thd3 :: (a, b, c) -> c

cataM :: (Monad m, Traversable (Base t), Recursive t) => (Base t a -> m a) -> t -> m a

Debugging/logging:

dbgLvl :: Int

Debugging flag shared by all modules. This is activated by setting the environment variable DEBUG=1..5

dbgPrint :: Int -> String -> IO ()

Print if the debug level is at or above a threshold.

dbgPrintLn :: Int -> String -> IO ()

dbgTrace :: Int -> String -> a -> a

Conditional version of Debug.Trace.trace

dbgTraceIt :: String -> a -> a

Yo, trace this msg.

minChatLvl :: Int

We should not create chatter below this level. DEBUG=1 is used for assertions only, not chatter.

internalError :: HasCallStack => String -> a

An error that is OUR FAULT, i.e. an internal bug in the compiler.

dumpIfSet :: Config -> DebugFlag -> String -> IO ()

Dump some output if the flag is set. Otherwise, do nothing.

Establish conventions for the output of #lang gibbon:

truePrinted :: String

For now this is designed to match the Racket output of "#lang gibbon" which itself should change once we implement a custom printer.

falsePrinted :: String

Orphan instances

(Out k, Out v) => Out (Map k v)

Instance details

Methods

docPrec :: Int -> Map k v -> Doc

doc :: Map k v -> Doc

docList :: [Map k v] -> Doc