Gibbon.GenericOps

Description

Generic operations that are supported by the various intermediate representations inside the compiler.

Documentation

class Interp a where

Pure Gibbon programs, at any stage of compilation, should always be evaluatable to a unique value. The only side effects are timing.

Minimal complete definition

interpNoLogs, interpWithStdout

Methods

interpNoLogs :: RunConfig -> a -> String

Interpret while ignoring timing constructs, and dropping the corresponding output to stdout.

interpWithStdout :: RunConfig -> a -> IO (String, [String])

Interpret and produce a "log" of output lines, as well as a final, printed result. The output lines include timing information.

Free Variables, generic interface

class FreeVars a where

Expression and program types which support a notion of free variables.

Minimal complete definition

gFreeVars

Methods

gFreeVars :: a -> Set Var

Return a set of free TERM variables. Does not return location variables.

Genereric interface for expressions and select passes

class (Show e, Out e, FreeVars e) => Expression e where

A generic interface to expressions found in different phases of the compiler.

Minimal complete definition

isTrivial

Associated Types

type TyOf e

The type representation used in this expression.

type LocOf e

The location (variable) representation used in this expression.

Methods

isTrivial :: e -> Bool

Is an expression considered trivial (duplicatable by the compiler)?

data NoExt l d

An uninhabidited type indicating that the base grammar is not extended with any additional constructs.

class Expression e => Flattenable e where

IRs amenable to flattening

Minimal complete definition

gFlattenExp, gFlattenGatherBinds

Methods

gFlattenExp :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> SyM e

Process an expression into a fully-flattened expression which typically includes a larger number of temporary, local variable bindings.

gFlattenGatherBinds :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> SyM ([Binds e], e)

A private method. Gather the bindings from a subexpression, but do not "discharge" them by creating a let expression. They are in order, so later may depend on earlier.

class Expression e => Simplifiable e where

IRs amenable to simplification/inlineTrivs

Minimal complete definition

gInlineTrivExp

Methods

gInlineTrivExp :: DDefs (TyOf e) -> e -> e

class Expression e => Typeable e where

This is NOT a replacement for any typechecker. This is supposed to be used just to recover the type of an expression in a type-environment Without this, we cannot have truly generic implementation of the Flattenable class, since we need to know the type of an expression before we discharge it with a LetE

Minimal complete definition

gTypeExp

Methods

gTypeExp :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> TyOf e