Gibbon.L0.Typecheck

Documentation

newtype TcM a

Constructors

TcM (ExceptT Doc PassM a)

runTcM :: TcM a -> PassM (Either Doc a)

err :: Doc -> TcM a

tcProg :: Prog0 -> PassM Prog0

tcFun :: DDefs0 -> Gamma -> FunDef0 -> PassM FunDef0

tcExps :: DDefs0 -> Subst -> Gamma -> Gamma -> [TyVar] -> [L Exp0] -> TcM (Subst, [Ty0], [L Exp0])

tcExp :: DDefs0 -> Subst -> Gamma -> Gamma -> [TyVar] -> L Exp0 -> TcM (Subst, Ty0, L Exp0)

tcCases :: DDefs0 -> Subst -> Gamma -> Gamma -> [TyVar] -> DDef0 -> [(DataCon, [(Var, Ty0)], L Exp0)] -> L Exp0 -> TcM (Subst, Ty0, [(DataCon, [(Var, Ty0)], L Exp0)])

instantiate :: TyScheme -> TcM ([Ty0], Ty0)

Instantiate the topmost for-alls of the argument type with meta type variables.

generalize :: Gamma -> Subst -> Ty0 -> TcM (Subst, TyScheme)

instDataConTy :: DDefs0 -> DataCon -> TcM ([Ty0], [Ty0], Ty0)

type Gamma = TyEnv TyScheme

newtype Subst

Constructors

Subst (Map MetaTv Ty0)

emptySubst :: Subst

zonkTy :: Subst -> Ty0 -> Ty0

Perform substitutions in the type.

zonkTyScheme :: Subst -> TyScheme -> TyScheme

zonkTyEnv :: Subst -> Gamma -> Gamma

zonkExp :: Subst -> L Exp0 -> L Exp0

substTyVarDDef :: DDef0 -> [Ty0] -> TcM DDef0

substTyVarExp :: Map TyVar Ty0 -> L Exp0 -> L Exp0

substTyVar :: Map TyVar Ty0 -> Ty0 -> Ty0

Replace the specified quantified type variables by given meta type variables.

tyVarToMetaTyl :: [Ty0] -> TcM (Map TyVar Ty0, [Ty0])

tyVarToMetaTy :: Ty0 -> TcM (Map TyVar Ty0, Ty0)

Replace the specified quantified type variables by given meta type variables.

unify :: L Exp0 -> Ty0 -> Ty0 -> TcM Subst

unifyl :: L Exp0 -> [Ty0] -> [Ty0] -> TcM Subst

unifyVar :: L Exp0 -> MetaTv -> Ty0 -> TcM Subst

occursCheck :: MetaTv -> Ty0 -> Bool

ensureEqualTy :: L Exp0 -> Ty0 -> Ty0 -> TcM ()

Orphan instances

FreeVars a => FreeVars (TyEnv a)

Instance details

Methods

gFreeVars :: TyEnv a -> Set Var