Gibbon.L1.Syntax

Description

The source language for recursive tree traversals. This is a first-order language for the "closed world" scenario: not integrating with a functional host language, but rather genarating C code like a DSL.

Generic types shared across all the IR's

data Prog ex

Complete programs include datatype definitions:

For evaluating a complete program, main's type will be an Int or a datatype. For running a pass benchmark, main will be Nothing and we will expect a "benchmark" function definition which consumes an appropriate packed AST datatype.

Constructors

Prog

Fields

ddefs :: DDefs (TyOf ex)

fundefs :: FunDefs ex

mainExp :: Maybe (ex, TyOf ex)

data FunDef ex

A function definiton indexed by a type and expression.

Constructors

FunDef

Fields

funName :: Var

funArg :: Var

funTy :: ArrowTy (TyOf ex)

funBody :: ex

type FunDefs ex = Map Var (FunDef ex)

A set of top-level recursive function definitions.

data PreExp (ext :: * -> * -> *) loc dec

The source language. It has pointer-based sums and products, as well as packed algebraic datatypes.

It is parameterized by an a potential extension point.

It is parameterized by loc, the type of locations.

It is parameterized by a decoration, d, attached to every binder.

Constructors

VarE Var
Variable reference

LitE Int
Numeric literal

LitSymE Var
A quoted symbol literal.

AppE Var [loc] (L (PreExp ext loc dec))
Apply a top-level / first-order function. Instantiate its type schema by providing location-variable arguments, if applicable.

PrimAppE (Prim dec) [L (PreExp ext loc dec)]
Primitive applications don't manipulate locations.

LetE (Var, [loc], dec, L (PreExp ext loc dec)) (L (PreExp ext loc dec))
One binding at a time. Allows binding a list of implicit *location* return vales from the RHS, plus a single "real" value. This list of implicit returnsb

IfE (L (PreExp ext loc dec)) (L (PreExp ext loc dec)) (L (PreExp ext loc dec))

MkProdE [L (PreExp ext loc dec)]
Tuple construction

ProjE Int (L (PreExp ext loc dec))
Tuple projection.

CaseE (L (PreExp ext loc dec)) [(DataCon, [(Var, loc)], L (PreExp ext loc dec))]
Case on a datatype. Each bound, unpacked variable lives at a fixed, read-only location.

DataConE loc DataCon [L (PreExp ext loc dec)]
Construct data that may unpack some fields. The location argument, if applicable, is the byte location at which to write the tag for the sum type.

TimeIt (L (PreExp ext loc dec)) dec Bool
The boolean being true indicates this TimeIt is really (iterate _) This iterate form is used for criterion-style benchmarking.

MapE (Var, Ty1, L (PreExp ext loc dec)) (L (PreExp ext loc dec))

FoldE

Fields

initial :: (Var, dec, L (PreExp ext loc dec))

iterator :: (Var, dec, L (PreExp ext loc dec))

body :: L (PreExp ext loc dec)

Ext (ext loc dec)
Extension point for downstream language extensions.

data Prim ty

Some of these primitives are (temporarily) tagged directly with their return types.

Constructors

AddP

SubP

MulP
May need more numeric primitives...

DivP

ModP
Integer division and modulus

EqSymP
Equality on Sym

EqIntP
Equality on Int

LtP

GtP
(and () for Int's

SymAppend
A quick hack till we have deterministic gensym

DictInsertP ty
takes dict, k,v; annotated with element type

DictLookupP ty
takes dict,k errors if absent; annotated with element type

DictEmptyP ty
annotated with element type to avoid ambiguity

DictHasKeyP ty
takes dict,k; returns a Bool, annotated with element type

ErrorP String ty
crash and issue a static error message. To avoid needing inference, this is labeled with a return type.

SizeParam

MkTrue
Zero argument constructor.

MkFalse
Zero argument constructor.

ReadPackedFile (Maybe FilePath) TyCon ty
Read (mmap) a binary file containing packed data. This must be annotated with the type of the file being read. The Ty tracks the type as the program evolvels (first PackedTy then CursorTy). The TyCon tracks the original type name.

PEndOf
This shouldn't be here. But this is the fastest way to encode indirection nodes right now. This can be an L2 extension, after we make InferLayout the pass that takes L1->L2, and InferLocations is an L2->L2 pass.

data UrTy a

Types include boxed/pointer-based products as well as unpacked algebraic datatypes. This data is parameterized to allow annotation on Packed types later on.

Constructors

IntTy

BoolTy

ProdTy [UrTy a]
An N-ary tuple

SymDictTy (UrTy a)
A map from SymTy to Ty ^ We allow built-in dictionaries from symbols to a value type.

PackedTy TyCon a
No type arguments to TyCons for now. (No polymorphism.)

ListTy (UrTy a)
These are not fully first class. They are only allowed as the fields of data constructors.

PtrTy
A machine pointer to a complete value in memory. This is decorated with the region it points into, which may affect the memory layout.

CursorTy
A cursor for reading or writing, which may point to an unkwown type or to a fraction of a complete value. It is a machine pointer that can point to any byte.

type family ArrowTy ty

A type family describing function types.

Core types specific to L1

type Prog1 = Prog (L Exp1)

An L1 program.

type FunDef1 = FunDef (L Exp1)

Function definition used in L1 programs.

type Exp1 = PreExp NoExt () Ty1

A convenient, default instantiation of the L1 expression type.

type Ty1 = UrTy ()

The type rperesentation used in L1.

pattern SymTy :: forall a. UrTy a

Interpreter

class Expression e => InterpE e where

Minimal complete definition

interpE

Associated Types

type ExpTy e

A temporary HACK which gives the type of the expression being interpreted. We need the type to get the proper FunDef, which is used to interpret function applications.

For PreExp or other datatypes which are actual expressions, this is set to the same datatype. However, when we're defining an instance of this class for an extension, it isn't clear how to infer the type of the expression without this.

Methods

interpE :: RunConfig -> DDefs (TyOf e) -> Map Var (FunDef (ExpTy e)) -> e -> WriterT Log (StateT Store IO) Value

Interpret an expression and return a Value

Functions on expressions

insertFD :: FunDef ex -> FunDefs ex -> FunDefs ex

Insert a FunDef into FunDefs. Raise an error if a function with the same name already exists.

fromListFD :: [FunDef ex] -> FunDefs ex

mapExt :: (e1 l d -> e2 l d) -> PreExp e1 l d -> PreExp e2 l d

Apply a function to the extension points only.

mapLocs :: (e l2 d -> e l2 d) -> PreExp e l2 d -> PreExp e l2 d

Apply a function to the locations only.

mapExprs :: (L Exp1 -> L Exp1) -> Prog1 -> Prog1

Transform the expressions within a program.

visitExp :: forall l1 l2 e1 e2 d1 d2. (l1 -> l2) -> (e1 l1 d1 -> e2 l2 d2) -> (d1 -> d2) -> PreExp e1 l1 d1 -> PreExp e2 l2 d2

Apply a function to the locations, extensions, and binder-decorations, respectively.

progToEnv :: Prog1 -> Env2 Ty1

Abstract some of the differences of top level program types, by having a common way to extract an initial environment.

getFunTy :: Var -> Prog ex -> ArrowTy (TyOf ex)

Look up the input/output type of a top-level function binding.

subst :: (Eq d, Eq l, Eq (e l d)) => Var -> L (PreExp e l d) -> L (PreExp e l d) -> L (PreExp e l d)

Substitute an expression in place of a variable.

substE :: (Eq d, Eq l, Eq (e l d)) => L (PreExp e l d) -> L (PreExp e l d) -> L (PreExp e l d) -> L (PreExp e l d)

Expensive subst that looks for a whole matching sub-EXPRESSION. If the old expression is a variable, this still avoids going under binder.

hasTimeIt :: L (PreExp e l d) -> Bool

Does the expression contain a TimeIt form?

projNonFirst :: (Out l, Out d, Out (e l d)) => Int -> L (PreExp e l d) -> L (PreExp e l d)

Project something which had better not be the first thing in a tuple.

mkProj :: (Eq a, Num a) => Int -> a -> L (PreExp e l d) -> L (PreExp e l d)

Project position K of N, unless (K,N) = (0,1) in which case no projection is necessary.

mkProd :: [L (PreExp e l d)] -> L (PreExp e l d)

Make a product type while avoiding unary products.

mkLets :: [(Var, [loc], dec, L (PreExp ext loc dec))] -> L (PreExp ext loc dec) -> L (PreExp ext loc dec)

Make a nested series of lets.

flatLets :: [(Var, [l], d, L (PreExp e l d))] -> L (PreExp e l d) -> L (PreExp e l d)

Alternative version of L1.mkLets that also flattens

Functions on types

mkProdTy :: [UrTy a] -> UrTy a

Same as mkProd, at the type level

projTy :: Out a => Int -> UrTy a -> UrTy a

voidTy :: Ty1

A makeshift void type.

isProdTy :: UrTy a -> Bool

Are values of this type tuples ?

isNestedProdTy :: UrTy a -> Bool

Do values of this type contain nested tuples ?

isPackedTy :: UrTy a -> Bool

Are values of this type Packed ?

hasPacked :: Show a => UrTy a -> Bool

Do values of this type contain packed data?

sizeOfTy :: UrTy a -> Maybe Int

Provide a size in bytes, if it is statically known.

primArgsTy :: Prim Ty1 -> [Ty1]

Type of the arguments for a primitive operation.

primRetTy :: Prim (UrTy a) -> UrTy a

Return type for a primitive operation.

dummyCursorTy :: UrTy a

Misc

numIndrsDataCon :: Out a => DDefs (UrTy a) -> DataCon -> Maybe Int

The number of indirections needed by a DataCon for full random access

assertTriv :: Expression e => L e -> a -> a

Ensure that an expression is trivial.

assertTrivs :: Expression e => [L e] -> a -> a

List version of assertTriv.