Gibbon.L2.Interp

Description

Interpreter reducing L2 programs to values.

Documentation

interp :: SizeEnv -> RunConfig -> ValEnv Exp2 -> DDefs Ty2 -> Map Var (FunDef Exp2) -> Exp2 -> InterpM Store Exp2 (Value Exp2, Size)

interpExt :: SizeEnv -> RunConfig -> ValEnv Exp2 -> DDefs Ty2 -> Map Var (FunDef Exp2) -> E2Ext LocVar Ty2 -> InterpM Store Exp2 (Value Exp2, Size)

clk :: Clock

strToInt :: String -> Int

newtype Store

A store is an address space full of buffers.

Constructors

Store (Map Var Buffer)

emptyStore :: Store

insertIntoStore :: Var -> Buffer -> InterpM Store Exp2 ()

lookupInStore :: Var -> InterpM Store Exp2 (Maybe Buffer)

lookupInStore' :: Var -> Store -> Maybe Buffer

newtype Buffer

Constructors

Buffer (Seq SerializedVal)

emptyBuffer :: Buffer

insertAtBuffer :: Int -> SerializedVal -> Buffer -> Buffer

Insert a value at a particular index in the buffer

dropInBuffer :: Int -> Buffer -> Buffer

data SerializedVal

Constructors

SerTag Word8 DataCon

SerInt Int

SerBool Int

SerFloat Double

SerPtr

Fields

bufID :: Var

offset :: Int

SerPad
Used to make values artificially occupy more cells in the buffer. For example, everywhere else we assume that Ints occupy 8 cells. But SerInt only occupies a single cell. To make SerInt occupy 8 cells, we add 7 SerPads after it.

byteSize :: SerializedVal -> Int

byteSizeOfTy :: UrTy d -> Maybe Int

Everything occupies a single cell in the Buffer.

deserialize :: Out ty => DDefs ty -> Store -> Buffer -> Value Exp2

Code to read a final answer back out.

data Size

Constructors

SOne Int

SMany [Size]

type SizeEnv = Map Var Size

sizeToInt :: Size -> Int

appendSize :: Size -> Size -> Size

Orphan instances

Out a => Out (Seq a)

Instance details

Methods

docPrec :: Int -> Seq a -> Doc

doc :: Seq a -> Doc

docList :: [Seq a] -> Doc