Gibbon.L2.Syntax

Description

An intermediate language with an effect system that captures traversals.

Extended language L2 with location types.

data E2Ext loc dec

The extension that turns L1 into L2.

Constructors

LetRegionE Region (L (E2 loc dec))
Not used until later on.

LetLocE loc (PreLocExp loc) (L (E2 loc dec))
Bind a new location.

RetE [loc] Var
Return a value together with extra loc values.

FromEndE loc
Bind a location from an EndOf location (for RouteEnds and after)

BoundsCheck Int loc loc
Bytes required, region, write cursor

IndirectionE TyCon DataCon (loc, Var) (loc, Var) (L (E2 loc dec))
An tagged indirection node

type Prog2 = Prog (L Exp2)

type FunDef2 = FunDef (L Exp2)

type FunDefs2 = FunDefs (L Exp2)

type Exp2 = E2 LocVar Ty2

Extended expressions, L2.

By adding a LocVar decoration, all data constructors, applications, and bindings gain a location annotation.

type E2 l d = PreExp E2Ext l d

Shorthand for recursions.

type Ty2 = UrTy LocVar

L1 Types extended with abstract Locations.

data Effect

The side-effect of evaluating a function.

Constructors

Traverse LocVar
The function, during its execution, traverses all of the value living at this location.

data ArrowTy2

Our type for functions grows to include effects, and explicit universal quantification over location/region variables.

Constructors

ArrowTy2

Fields

locVars :: [LRM]

Universally-quantified location params. Only these should be referenced in arrIn/arrOut.

arrIn :: Ty2

Input type for the function.

arrEffs :: Set Effect

These are present-but-empty initially, and the populated by InferEffects.

arrOut :: Ty2

Output type for the function.

locRets :: [LocRet]

L2B feature: multi-valued returns.

data LocRet

Locations (end-witnesses) returned from functions after RouteEnds.

Constructors

EndOf LRM

type LocExp = PreLocExp LocVar

data PreLocExp loc

Define a location in terms of a different location.

Constructors

StartOfLE Region

AfterConstantLE Int loc

AfterVariableLE Var loc

InRegionLE Region

FromEndLE loc

Regions and locations

type LocVar = Var

Abstract location variables.

data Region

An abstract region identifier. This is used inside type signatures and elsewhere.

Constructors

GlobR Var Multiplicity
A global region with lifetime equal to the whole program.

DynR Var Multiplicity
A dynamic region that may be created or destroyed, tagged by an identifier.

VarR Var
A region metavariable that can range over either global or dynamic regions.

MMapR Var
A region that doesn't result in an (explicit) memory allocation. It merely ensures that there are no free locations in the program.

data Modality

The modality of locations and cursors: input/output, for reading and writing, respectively.

Constructors

Input

Output

data LRM

A location and region, together with modality.

Constructors

LRM

Fields

lrmLoc :: LocVar

lrmReg :: Region

lrmMode :: Modality

dummyLRM :: LRM

A designated doesn't-really-exist-anywhere location.

data Multiplicity

Region variants (multiplicities)

Constructors

Bounded
Contain a finite number of values and can be stack-allocated.

Infinite
Consist of a linked list of buffers, spread throughout memory (though possible constrained to 4GB regions). Writing into these regions requires bounds-checking. The buffers can start very small at the head of the list, but probably grow geometrically in size, making the cost of traversing all of them logarithmic.

BigInfinite
These regions are infinite, but also have the expectation of containing many values. Thus we give them large initial page sizes. This is also could be the appropriate place to use mmap to grow the region and to establish guard places.

regionToVar :: Region -> Var

Operations on types

allLocVars :: ArrowTy2 -> [LocVar]

Retrieve all LocVars from a fn type (Arrow)

inLocVars :: ArrowTy2 -> [LocVar]

outLocVars :: ArrowTy2 -> [LocVar]

outRegVars :: ArrowTy2 -> [LocVar]

inRegVars :: ArrowTy2 -> [LocVar]

substLoc :: Map LocVar LocVar -> Ty2 -> Ty2

Apply a location substitution to a type.

substLoc' :: LocVar -> Ty2 -> Ty2

Like substLoc, but constructs the map for you..

substLocs :: Map LocVar LocVar -> [Ty2] -> [Ty2]

List version of substLoc.

substLocs' :: [LocVar] -> [Ty2] -> [Ty2]

Like substLocs, but constructs the map for you.

substEffs :: Map LocVar LocVar -> Set Effect -> Set Effect

Apply a substitution to an effect set.

prependArgs :: [UrTy l] -> UrTy l -> UrTy l

Injected cursor args go first in input and output:

stripTyLocs :: UrTy a -> UrTy ()

Remove the extra location annotations.

locsInTy :: Ty2 -> [LocVar]

Collect all the locations mentioned in a type.

Other helpers

revertToL1 :: Prog2 -> Prog1

occurs :: Set Var -> L Exp2 -> Bool

Does a variable occur in an expression ?

mapPacked :: (Var -> l -> UrTy l) -> UrTy l -> UrTy l

depList :: L Exp2 -> [(Var, Var, [Var])]

Build a dependency list which can be later converted to a graph

module Gibbon.Language

Orphan instances

FunctionTy Ty2
Function types know about locations and traversal effects.

Instance details

Associated Types

type ArrowTy Ty2 :: *

Methods

inTy :: ArrowTy Ty2 -> Ty2

outTy :: ArrowTy Ty2 -> Ty2

Out a => Out (Set a)

Instance details

Methods

docPrec :: Int -> Set a -> Doc

doc :: Set a -> Doc

docList :: [Set a] -> Doc