Gibbon.L3.Syntax

Description

An intermediate language which makes cursors explicit

Extended language

data E3Ext loc dec

The extension that turns L1 into L3.

Constructors

ReadInt Var
One cursor in, (int, cursor') out

WriteInt Var (L (PreExp E3Ext loc dec))
Write int at cursor, and return a cursor

AddCursor Var (L (PreExp E3Ext loc dec))
Add a constant offset to a cursor variable

ReadTag Var
One cursor in, (tag,cursor) out

WriteTag DataCon Var
Write Tag at Cursor, and return a cursor

NewBuffer Multiplicity
Create a new buffer, and return a cursor

ScopedBuffer Multiplicity
Create a temporary scoped buffer, and return a cursor

InitSizeOfBuffer Multiplicity
Returns the initial buffer size for a specific multiplicity

SizeOfPacked Var Var
Takes in start and end cursors, and returns an Int we'll probably represent (sizeof x) as (end_x - start_x) / INT

SizeOfScalar Var
sizeof(var)

BoundsCheck Int Var Var
Bytes required, region, write cursor

ReadCursor Var
Reads and returns the cursor at Var

WriteCursor Var (L (PreExp E3Ext loc dec))
Write a cursor, and return a cursor

BumpRefCount Var Var
Given an end-of-region ptr, bump it's refcount. Return the updated count (optional).

type Prog3 = Prog (L Exp3)

type FunDef3 = FunDef (L Exp3)

type FunDefs3 = FunDefs (L Exp3)

type Exp3 = PreExp E3Ext () Ty3

type Ty3 = UrTy ()

Functions

eraseLocMarkers :: DDef Ty2 -> DDef Ty3

Erase LocVar markers from the data definition

mapMExprs :: Monad m => (Env2 Ty3 -> L Exp3 -> m (L Exp3)) -> Prog3 -> m Prog3

Map exprs with an initial type environment: Exactly the same function that was in L2 before

cursorizeTy :: UrTy a -> UrTy b

toL3Prim :: Prim Ty2 -> Prim Ty3

progToEnv :: Prog3 -> Env2 Ty3

Abstract some of the differences of top level program types, by having a common way to extract an initial environment.