Gibbon.L4.Syntax

Description

Defines the target language for first-order L1 language with C code generator for it.

Documentation

data Var

type Tag = Word8

data Tail

Constructors

RetValsT [Triv]
Only in tail position, for returning from a function.

AssnValsT
INTERNAL ONLY: used for assigning instead of returning.

Fields

upd :: [(Var, Ty, Triv)]

bod_maybe :: Maybe Tail

LetCallT

Fields

async :: Bool

Whether this call should be executed asynchronously (cilk_spawn).

binds :: [(Var, Ty)]

rator :: Var

rands :: [Triv]

bod :: Tail

LetPrimCallT

Fields

binds :: [(Var, Ty)]

prim :: Prim

rands :: [Triv]

bod :: Tail

LetTrivT

Fields

bnd :: (Var, Ty, Triv)

bod :: Tail

LetIfT

Fields

binds :: [(Var, Ty)]

ife :: (Triv, Tail, Tail)

bod :: Tail

LetUnpackT
Unpack a struct pointer (variable of PtrTy) into local fields.

Fields

binds :: [(Var, Ty)]

ptr :: Var

Var pointing to a PtrTy

bod :: Tail

LetAllocT
Allocate storage for a struct of the given type, Initialize all fields Return PtrTy.

Fields

lhs :: Var

Var to bind to PtrTy

vals :: [(Ty, Triv)]

Fields of allocated struct

bod :: Tail

IfT

Fields

tst :: Triv

con :: Tail

els :: Tail

ErrT String

LetTimedT
This is like a one-armed if. It needs a struct return.

Fields

isIter :: Bool

Run the RHS multiple times, if true.

binds :: [(Var, Ty)]

timed :: Tail

bod :: Tail

Switch Label Triv Alts (Maybe Tail)
For casing on numeric tags or integers.

TailCall Var [Triv]

Goto Label

data Triv

Constructors

VarTriv Var

IntTriv Int64

TagTriv Tag

data Ty

Constructors

IntTy
8 byte integers.

BoolTy
1 byte integers.

TagTyPacked
A single byte / Word8. Used in PACKED mode.

TagTyBoxed
A tag used in the UNPACKED, boxed, pointer-based, graph-of-structs representation. This can usually be the same as TagTy, but needn't necessarily be.

SymTy
Symbols used in writing compiler passes. It's an alias for Int, an index into a symbol table.

CursorTy
A byte-indexing pointer. This is always a pointer to a raw buffer of bytes that does not contain pointers.

PtrTy
A machine word. Same width as IntTy. Untyped. This is a pointer to a struct value which may contain other pointers.

RegionTy
Region start and a refcount

ChunkTy
Start and end pointers

ProdTy [Ty]

SymDictTy Ty
We allow built-in dictionaries from symbols to a value type.

data Prim

Constructors

AddP

SubP

MulP

DivP

ModP

EqP

LtP

GtP

LtEqP

GtEqP

OrP

AndP

ExpP

RandP

SizeParam

DictInsertP Ty
takes k,v,dict

DictLookupP Ty
takes k,dict, errors if absent

DictEmptyP Ty

DictHasKeyP Ty

ReadPackedFile (Maybe FilePath) TyCon

NewBuffer Multiplicity
Allocate a new buffer, return a cursor.

ScopedBuffer Multiplicity
Returns a pointer to a buffer, with the invariant that data written to this region is no longer used after the enclosing function returns. I.e. this can be stack allocated data.

InitSizeOfBuffer Multiplicity
Returns the initial buffer size for a specific multiplicity

MMapFileSize Var

WriteTag
Write a static tag value, takes a cursor to target.

WriteInt
Write (leaf) data, takes cursor,int

ReadTag
Read one byte from the cursor and advance it.

ReadInt
Read an 8 byte Int from the cursor and advance.

ReadCursor
Read and return a cursor

WriteCursor

ReadBool

WriteBool
Read / write 1 byte integers, and advance.

BoundsCheck

BumpRefCount

FreeBuffer

SizeOfPacked
Take start and end cursors and return size of data they represent This could be represented as (end - start) / (sizeof(Int))

SizeOfScalar
Takes in a variable, and returns an int, sizeof(var)

GetFirstWord
takes a PtrTy, returns IntTy containing the (first) word pointed to.

PrintInt
Print an integer to stdout.

PrintString String
Print a constant string to stdout. TODO: add string values to the language.

ParSync
cilk_sync

data FunDecl

Constructors

FunDecl

Fields

funName :: Var

funArgs :: [(Var, Ty)]

funRetTy :: Ty

funBody :: Tail

isPure :: Bool

data Alts

Switch alternatives.

Constructors

TagAlts [(Tag, Tail)]
Casing on tags.

IntAlts [(Int64, Tail)]
Casing on integers.

data Prog

Constructors

Prog

Fields

fundefs :: [FunDecl]

mainExp :: Maybe MainExp

data MainExp

Constructors

PrintExp Tail
Evaluate the expression and print the result. Type of the expression must will eventually be anything, but not all types support printing currently [2017.01.03].

type Label = Var

Utility functions

withTail :: MonadState Int m => (Tail, Ty) -> ([Triv] -> Tail) -> m Tail

Extend the tail of a Tail. Take the return values from a Tail expression and do some more computation.

WARNING: presently this may invoke the given function more than once and duplicate code.

fromL3Ty :: Ty3 -> Ty

voidTy :: Ty

Orphan instances

Out Int64

Instance details

Methods

docPrec :: Int -> Int64 -> Doc

doc :: Int64 -> Doc

docList :: [Int64] -> Doc

Out Word8

Instance details

Methods

docPrec :: Int -> Word8 -> Doc

doc :: Word8 -> Doc

docList :: [Word8] -> Doc