Gibbon.Language.Constants

Documentation

redirectionSize :: Int

redirectionTag :: DataCon

isRedirectionTag :: DataCon -> Bool

redirectionAlt :: Num a => a

indirectionTag :: DataCon

isIndirectionTag :: DataCon -> Bool

indirectionAlt :: Num a => a

toAbsRANDataCon :: DataCon -> DataCon

isAbsRANDataCon :: DataCon -> Bool

toRelRANDataCon :: DataCon -> DataCon

isRelRANDataCon :: DataCon -> Bool

fromRANDataCon :: DataCon -> DataCon

mkUnpackerName :: TyCon -> Var

Map a DataCon onto the name of the generated unpack function.

isUnpackerName :: Var -> Bool

mkPrinterName :: TyCon -> Var

Map a DataCon onto the name of the generated print function.

isPrinterName :: Var -> Bool

mkCopyFunName :: TyCon -> Var

isCopyFunName :: Var -> Bool

mkCopySansPtrsFunName :: TyCon -> Var

isCopySansPtrsFunName :: Var -> Bool

mkTravFunName :: TyCon -> Var

isTravFunName :: Var -> Bool

mkRelOffsetsFunName :: DataCon -> Var

isRelOffsetsFunName :: Var -> Bool