Gibbon.Language.Syntax

Datatype definitions

type DDefs a = Map Var (DDef a)

type DataCon = String

type TyCon = String

type Tag = Word8

type IsBoxed = Bool

data DDef a

Data type definitions.

Monomorphism: In the extreme case we can strip packed datatypes of all type parameters, or we can allow them to retain type params but require that they always be fully instantiated to monomorphic types in the context of our monomorphic programs.

Here we allow individual to be marked with whether or not they should be boxed. We say that a regular, pointer-based datatype has all-boxed fields, whereas a fully serialized datatype has no boxed fields.

Constructors

DDef

Fields

tyName :: Var

tyArgs :: [TyVar]

dataCons :: [(DataCon, [(IsBoxed, a)])]

lookupDDef :: Out a => DDefs a -> TyCon -> DDef a

Lookup a ddef in its entirety

getConOrdering :: Out a => DDefs a -> TyCon -> [DataCon]

Get the canonical ordering for data constructors, currently based on ordering in the original source code. Takes a TyCon as argument.

getTyOfDataCon :: Out a => DDefs a -> DataCon -> TyCon

Lookup the name of the TyCon that goes with a given DataCon. Must be unique!

lookupDataCon :: Out a => DDefs a -> DataCon -> [a]

Lookup the arguments to a data contstructor.

lkp :: Out a => DDefs a -> DataCon -> (Var, (DataCon, [(IsBoxed, a)]))

Lookup a Datacon. Return (TyCon, (DataCon, [flds]))

lookupDataCon' :: Out a => DDef a -> DataCon -> [a]

Like lookupDataCon but lookup arguments to a data contstructor for a specific instance of a datatype.

lookupDataCon' (Maybe Int) Just = [Int]

insertDD :: DDef a -> DDefs a -> DDefs a

emptyDD :: DDefs a

fromListDD :: [DDef a] -> DDefs a

isVoidDDef :: DDef a -> Bool

Is this an empty type (like 'data Void' in Haskell) ?

Function definitions

class (Out (ArrowTy ty), Show (ArrowTy ty)) => FunctionTy ty where

A type family describing function types.

Associated Types

type ArrowTy ty

Methods

inTys :: ArrowTy ty -> [ty]

outTy :: ArrowTy ty -> ty

type FunDefs ex = Map Var (FunDef ex)

A set of top-level recursive function definitions.

data FunDef ex

A function definiton indexed by a type and expression.

Constructors

FunDef

Fields

funName :: Var

funArgs :: [Var]

funTy :: ArrowTy (TyOf ex)

funBody :: ex

funRec :: FunRec

funInline :: FunInline

data FunRec

Constructors

Rec

NotRec

TailRec

data FunInline

Constructors

Inline

NoInline

Inlineable

insertFD :: FunDef ex -> FunDefs ex -> FunDefs ex

Insert a FunDef into FunDefs. Raise an error if a function with the same name already exists.

fromListFD :: [FunDef ex] -> FunDefs ex

initFunEnv :: FunDefs a -> TyEnv (ArrowTy (TyOf a))

Programs

data Prog ex

Complete programs include datatype definitions:

For evaluating a complete program, main's type will be an Int or a datatype. For running a pass benchmark, main will be Nothing and we will expect a "benchmark" function definition which consumes an appropriate packed AST datatype.

Constructors

Prog

Fields

ddefs :: DDefs (TyOf ex)

fundefs :: FunDefs ex

mainExp :: Maybe (ex, TyOf ex)

progToEnv :: Prog a -> Env2 (TyOf a)

Abstract some of the differences of top level program types, by having a common way to extract an initial environment. The initial environment has types only for functions.

getFunTy :: Var -> Prog ex -> ArrowTy (TyOf ex)

Look up the input/output type of a top-level function binding.

Environments

type TyEnv a = Map Var a

A simple type environment

data Env2 a

A common currency for a two part environment consisting of function bindings and regular value bindings.

Constructors

Env2

Fields

vEnv :: TyEnv a

fEnv :: TyEnv (ArrowTy a)

emptyEnv2 :: Env2 a

extendVEnv :: Var -> a -> Env2 a -> Env2 a

Extend non-function value environment.

extendsVEnv :: Map Var a -> Env2 a -> Env2 a

Extend multiple times in one go.

lookupVEnv :: Out a => Var -> Env2 a -> a

extendFEnv :: Var -> ArrowTy a -> Env2 a -> Env2 a

Extend function type environment.

lookupFEnv :: Out (ArrowTy a) => Var -> Env2 a -> ArrowTy a

Expresssions and thier types

data PreExp (ext :: Type -> Type -> Type) loc dec

The source language. It has pointer-based sums and products, as well as packed algebraic datatypes.

It is parameterized by an a potential extension point.

It is parameterized by loc, the type of locations.

It is parameterized by a decoration, d, attached to every binder.

Constructors

VarE Var
Variable reference

LitE Int
Numeric literal

FloatE Double
Floating point literal

LitSymE Var
A quoted symbol literal.

AppE Var [loc] [PreExp ext loc dec]
Apply a top-level / first-order function. Instantiate its type schema by providing location-variable arguments, if applicable.

PrimAppE (Prim dec) [PreExp ext loc dec]
Primitive applications don't manipulate locations.

LetE (Var, [loc], dec, PreExp ext loc dec) (PreExp ext loc dec)
One binding at a time. Allows binding a list of implicit *location* return vales from the RHS, plus a single "real" value. This list of implicit returnsb

IfE (PreExp ext loc dec) (PreExp ext loc dec) (PreExp ext loc dec)

MkProdE [PreExp ext loc dec]
Tuple construction

ProjE Int (PreExp ext loc dec)
Tuple projection.

CaseE (PreExp ext loc dec) [(DataCon, [(Var, loc)], PreExp ext loc dec)]
Case on a datatype. Each bound, unpacked variable lives at a fixed, read-only location.

DataConE loc DataCon [PreExp ext loc dec]
Construct data that may unpack some fields. The location argument, if applicable, is the byte location at which to write the tag for the sum type.

TimeIt (PreExp ext loc dec) dec Bool
The boolean being true indicates this TimeIt is really (iterate _) This iterate form is used for criterion-style benchmarking.

WithArenaE Var (PreExp ext loc dec)

SpawnE Var [loc] [PreExp ext loc dec]

SyncE

MapE (Var, dec, PreExp ext loc dec) (PreExp ext loc dec)

FoldE

Fields

initial :: (Var, dec, PreExp ext loc dec)

iterator :: (Var, dec, PreExp ext loc dec)

body :: PreExp ext loc dec

Ext (ext loc dec)
Extension point for downstream language extensions.

data Prim ty

Some of these primitives are (temporarily) tagged directly with their return types.

Constructors

AddP

SubP

MulP
May need more numeric primitives...

DivP

ModP
Integer division and modulus

ExpP
Exponentiation

RandP
Generate a random number. Translates to 'rand()' in C.

EqIntP
Equality on Int

LtP

GtP
(and () for Int's

LtEqP

GtEqP
and=

FAddP

FSubP

FMulP

FDivP

FExpP

FRandP

EqFloatP

FLtP

FGtP

FLtEqP

FGtEqP

FSqrtP

IntToFloatP

FloatToIntP

FTanP
Translates to 'tan()' in C.

EqSymP
Equality on Sym

EqBenchProgP String

OrP

AndP

MkTrue
Zero argument constructor.

MkFalse
Zero argument constructor.

ErrorP String ty
crash and issue a static error message. To avoid needing inference, this is labeled with a return type.

SizeParam

IsBig
Check the size of constructors with size.

GetNumProcessors
Return the number of processors

PrintInt
Print an integer to standard out

PrintFloat
Print a floating point number to standard out

PrintBool
Print a boolean to standard out

PrintSym
Print a symbol to standard out

ReadInt
Read an int from standard in

DictInsertP ty
takes dict, k,v; annotated with element type

DictLookupP ty
takes dict,k errors if absent; annotated with element type

DictEmptyP ty
annotated with element type to avoid ambiguity

DictHasKeyP ty
takes dict,k; returns a Bool, annotated with element type

SymSetEmpty
Creates an empty set

SymSetInsert
Inserts a symbol into a set of symbols

SymSetContains
Queries if a symbol is in a set

SymHashEmpty
Create empty hash table of symbols

SymHashInsert
Insert a symbol into a hash table

SymHashLookup
Look up a symbol in a hash table (takes default symbol)

SymHashContains
Queries if a symbol is in a hash

IntHashEmpty
Create empty hash table of integers

IntHashInsert
Insert an integer into a hash table

IntHashLookup
Look up a integer in a hash table (takes default integer)

PDictAllocP ty ty
annotated with element type to avoid ambiguity

PDictInsertP ty ty
takes dict, k, v; annotated with element type

PDictLookupP ty ty
takes dict, k. errors if absent; annotated with element type

PDictHasKeyP ty ty
takes dict,k; returns a Bool, annotated with element type

PDictForkP ty ty
takes dict; returns thread safe safe dicts.

PDictJoinP ty ty
takes 2 dicts; returns a merged dict.

LLAllocP ty

LLIsEmptyP ty

LLConsP ty

LLHeadP ty

LLTailP ty

LLFreeP ty
Free the list, and it's data.

LLFree2P ty
Free list struct, but not it's data.

LLCopyP ty
Copy the list node.

VAllocP ty
Allocate a vector

VFreeP ty
Free a vector, and it's data.

VFree2P ty
Free the vector struct, but not it's data.

VLengthP ty
Length of the vector

VNthP ty
Fetch the nth element

VSliceP ty
An efficient slice operation

InplaceVUpdateP ty
Update ith element of the vector

VConcatP ty
Flatten a vector

VSortP ty
A sort primop that accepts a function pointer

InplaceVSortP ty
A sort primop that sorts the array in place

VMergeP ty
ASSUMPTION: the vectors being merged have the same underlying mutable array. This assumption is checked at the type level with a Rank-2 type variable. But this evidence is erased (by the desugarer) by the time we get to L0.

Write3dPpmFile FilePath

ReadPackedFile (Maybe FilePath) TyCon (Maybe Var) ty
Read (mmap) a binary file containing packed data. This must be annotated with the type of the file being read. The Ty tracks the type as the program evolvels (first PackedTy then CursorTy). The TyCon tracks the original type name. The variable represents the region that this file will be mapped to, and is set by InferLocations.

WritePackedFile FilePath ty
Write a packed value to a file. To enable re-reading this packed value with Gibbon, this primitive gets rid of any absolute pointers in the value. First, it inlines (by copying) any regions pointed to by the packed value. Next, random access nodes are eliminated. We could change them to relative pointers (numeric offsets), but for a first version we can simplify things by getting rid of them completely.

ReadArrayFile (Maybe (FilePath, Int)) ty
Parse a file into a Vector. This is decorated with the element type. If the element type is a struct, like (Int, Int) for example, each line must contain 2 numbers separated by a space. The Int is the number of lines in the file.

RequestEndOf
Conveys a demand for the "end of" some packed value, which is fulfilled by Cursorize. N.B. the argument must be a VarE that refers to a packed value.

RequestSizeOf
Like RequestEndOf but gets the size of a packed value. Assume that the value is written in a contiguous region, and size = end_v - v.

Gensym

data UrTy a

Types include boxed/pointer-based products as well as unpacked algebraic datatypes. This data is parameterized to allow annotation on Packed types later on.

Constructors

IntTy

FloatTy

SymTy
Symbols used in writing compiler passes.

BoolTy

ProdTy [UrTy a]
An N-ary tuple

SymDictTy (Maybe Var) (UrTy ())
A map from SymTy to Ty ^ We allow built-in dictionaries from symbols to a value type.

PackedTy TyCon a
No type arguments to TyCons for now. (No polymorphism.)

VectorTy (UrTy a)
Vectors are decorated with the types of their elements; which can only include scalars or flat products of scalars.

PDictTy (UrTy a) (UrTy a)
Thread safe dictionaries decorated with key and value type.

ListTy (UrTy a)
Linked lists are decorated with the types of their elements; which can only include scalars or flat products of scalars.

ArenaTy
Collection of allocated, non-packed values

SymSetTy
Set of symbols

SymHashTy
Hash table of symbols

IntHashTy
Hash table of integers

PtrTy
A machine pointer tvo a complete value in memory. This is decorated with the region it points into, which may affect the memory layout.

CursorTy
A cursor for reading or writing, which may point to an unkwown type or to a fraction of a complete value. It is a machine pointer that can point to any byte.

Functors for recursion-schemes

data PreExpF (ext :: Type -> Type -> Type) (loc :: Type) (dec :: Type) r

Constructors

VarEF Var

LitEF Int

FloatEF Double

LitSymEF Var

AppEF Var [loc] [r]

PrimAppEF (Prim dec) [r]

LetEF (Var, [loc], dec, r) r

IfEF r r r

MkProdEF [r]

ProjEF Int r

CaseEF r [([Char], [(Var, loc)], r)]

DataConEF loc [Char] [r]

TimeItF r dec Bool

WithArenaEF Var r

SpawnEF Var [loc] [r]

SyncEF

MapEF (Var, dec, r) r

FoldEF

Fields

initialF :: (Var, dec, r)

iteratorF :: (Var, dec, r)

bodyF :: r

ExtF (ext loc dec)

data PrimF (ty :: Type) r

Constructors

AddPF

SubPF

MulPF

DivPF

ModPF

ExpPF

RandPF

EqIntPF

LtPF

GtPF

LtEqPF

GtEqPF

FAddPF

FSubPF

FMulPF

FDivPF

FExpPF

FRandPF

EqFloatPF

FLtPF

FGtPF

FLtEqPF

FGtEqPF

FSqrtPF

IntToFloatPF

FloatToIntPF

FTanPF

EqSymPF

EqBenchProgPF [Char]

OrPF

AndPF

MkTrueF

MkFalseF

ErrorPF [Char] ty

SizeParamF

IsBigF

GetNumProcessorsF

PrintIntF

PrintFloatF

PrintBoolF

PrintSymF

ReadIntF

DictInsertPF ty

DictLookupPF ty

DictEmptyPF ty

DictHasKeyPF ty

SymSetEmptyF

SymSetInsertF

SymSetContainsF

SymHashEmptyF

SymHashInsertF

SymHashLookupF

SymHashContainsF

IntHashEmptyF

IntHashInsertF

IntHashLookupF

PDictAllocPF ty ty

PDictInsertPF ty ty

PDictLookupPF ty ty

PDictHasKeyPF ty ty

PDictForkPF ty ty

PDictJoinPF ty ty

LLAllocPF ty

LLIsEmptyPF ty

LLConsPF ty

LLHeadPF ty

LLTailPF ty

LLFreePF ty

LLFree2PF ty

LLCopyPF ty

VAllocPF ty

VFreePF ty

VFree2PF ty

VLengthPF ty

VNthPF ty

VSlicePF ty

InplaceVUpdatePF ty

VConcatPF ty

VSortPF ty

InplaceVSortPF ty

VMergePF ty

Write3dPpmFileF [Char]

ReadPackedFileF (Maybe [Char]) [Char] (Maybe Var) ty

WritePackedFileF [Char] ty

ReadArrayFileF (Maybe ([Char], Int)) ty

RequestEndOfF

RequestSizeOfF

GensymF

data UrTyF (a :: Type) r

Constructors

IntTyF

FloatTyF

SymTyF

BoolTyF

ProdTyF [r]

SymDictTyF (Maybe Var) (UrTy ())

PackedTyF [Char] a

VectorTyF r

PDictTyF r r

ListTyF r

ArenaTyF

SymSetTyF

SymHashTyF

IntHashTyF

PtrTyF

CursorTyF

Generic operations

class FreeVars a where

Expression and program types which support a notion of free variables.

Methods

gFreeVars :: a -> Set Var

Return a set of free TERM variables. Does not return location variables.

class (Show e, Out e, FreeVars e) => Expression e where

A generic interface to expressions found in different phases of the compiler.

Associated Types

type TyOf e

The type representation used in this expression.

type LocOf e

The location (variable) representation used in this expression.

Methods

isTrivial :: e -> Bool

Is an expression considered trivial (duplicatable by the compiler)?

type Binds e = (Var, [LocOf e], TyOf e, e)

class Expression e => Flattenable e where

IRs amenable to flattening

Methods

gFlattenExp :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> PassM e

Process an expression into a fully-flattened expression which typically includes a larger number of temporary, local variable bindings.

gFlattenGatherBinds :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> PassM ([Binds e], e)

A private method. Gather the bindings from a subexpression, but do not "discharge" them by creating a let expression. They are in order, so later may depend on earlier.

class Expression e => Simplifiable e where

IRs amenable to simplification/inlineTrivs. Note that there's a separate SimplifiableExt for simplifying extensions. Simplifiable is the only class which makes such a distinction -- b/c when it's simplifying an extension point, the type of the environment would still be 'M.Map Var e', where e is a top-level IR. Right now we don't have a class (and probably don't want to have one as well) which ties an extension point with an IR. Keeping these classes separate works out nicely.

Methods

gInlineTrivExp :: Map Var e -> e -> e

class Expression e => SimplifiableExt e ext where

Methods

gInlineTrivExt :: Map Var e -> ext -> ext

class Expression e => Typeable e where

This is NOT a replacement for any typechecker. This only recover type of an expression given a type-environment. Without this, we cannot have truly generic Flattenable, b/c we need to know the type of an expression before we bind it with a LetE.

Methods

gRecoverType :: DDefs (TyOf e) -> Env2 (TyOf e) -> e -> TyOf e

class Expression e => Substitutable e where

Generic substitution over expressions.

Methods

gSubst :: Var -> e -> e -> e

gSubstE :: e -> e -> e -> e

class Expression e => SubstitutableExt e ext where

Methods

gSubstExt :: Var -> e -> ext -> ext

gSubstEExt :: e -> e -> ext -> ext

class Renamable e where

Alpha renaming, without worrying about name capture -- assuming that Freshen has run before!

Methods

gRename :: Map Var Var -> e -> e

Helpers for writing instances

type HasSimplifiable e l d = (Show l, Out l, Show d, Out d, Expression (e l d), SimplifiableExt (PreExp e l d) (e l d))

type HasSimplifiableExt e l d = (Show l, Out l, Show d, Out d, Simplifiable (PreExp e l d))

type HasSubstitutable e l d = (Expression (e l d), SubstitutableExt (PreExp e l d) (e l d), Eq d, Show d, Out d, Eq l, Show l, Out l, Eq (e l d))

type HasSubstitutableExt e l d = (Eq d, Show d, Out d, Eq l, Show l, Out l, Substitutable (PreExp e l d))

type HasRenamable e l d = (Renamable l, Renamable d, Renamable (e l d))

type HasOut ex = (Out ex, Out (TyOf ex), Out (ArrowTy (TyOf ex)))

type HasShow ex = (Show ex, Show (TyOf ex), Show (ArrowTy (TyOf ex)))

type HasEq ex = (Eq ex, Eq (TyOf ex), Eq (ArrowTy (TyOf ex)))

type HasGeneric ex = (Generic ex, Generic (TyOf ex), Generic (ArrowTy (TyOf ex)))

type HasNFData ex = (NFData ex, NFData (TyOf ex), NFData (ArrowTy (TyOf ex)))

Interpreter

class Expression e => Interp s e where

Pure Gibbon programs, at any stage of compilation, should always be evaluatable to a unique value. The only side effects are timing.

Methods

gInterpExp :: RunConfig -> ValEnv e -> DDefs (TyOf e) -> FunDefs e -> e -> InterpM s e (Value e)

class (Expression e, Expression ext) => InterpExt s e ext where

Methods

gInterpExt :: RunConfig -> ValEnv e -> DDefs (TyOf e) -> FunDefs e -> ext -> InterpM s e (Value e)

class Interp s e => InterpProg s e where

Minimal complete definition

gInterpProg

Methods

gInterpProg :: s -> RunConfig -> Prog e -> IO (s, Value e, ByteString)

gInterpNoLogs :: s -> RunConfig -> Prog e -> String

Interpret while ignoring timing constructs, and dropping the corresponding output to stdout.

gInterpWithStdout :: s -> RunConfig -> Prog e -> IO (String, [String])

Interpret and produce a "log" of output lines, as well as a final, printed result. The output lines include timing information.

data Value e

It's a first order language with simple values.

Constructors

VInt Int

VFloat Double

VSym String

VBool Bool

VDict (Map (Value e) (Value e))

VProd [Value e]

VList [Value e]

VPacked DataCon [Value e]

VLoc

Fields

bufID :: Var

offset :: Int

VCursor

Fields

bufID :: Var

offset :: Int

VPtr
Cursor are a pointer into the Store plus an offset into the Buffer.

Fields

bufID :: Var

offset :: Int

VLam [Var] e (ValEnv e)

VWrapId Int (Value e)
A wrapper for vectors that wraps the value with an "id". All Inplace* operations use this "id" to update the value in ValEnv.

type ValEnv e = Map Var (Value e)

type InterpLog = Builder

data InterpM s e a

runInterpM :: InterpM s e a -> s -> IO (a, InterpLog, s)

execAndPrint :: InterpProg s ex => s -> RunConfig -> Prog ex -> IO ()