Gibbon.Passes.Flatten

Description

Put the program in A-normal form where only varrefs and literals are allowed in operand position.

Documentation

flattenL1 :: Prog -> SyM Prog

Flatten ensures that function operands are "trivial".

In the process, it also lifts lets out of case scrutinees, if conditions, and tuple operands.

Note that it does not require tail expressions to be trivial. For example, it allows AppE and PrimAppE in the body of a let-expression.

flattenL2 :: Flattenable (E2Ext Var (UrTy LocVar)) => Prog -> SyM Prog

flattenL3 :: Prog -> SyM Prog